% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amr_complexUpset.R
\name{amr_complexUpset}
\alias{amr_complexUpset}
\title{Create an Upset plot for AMR data}
\usage{
amr_complexUpset(
  binary_matrix,
  min_set_size = 10,
  mic_disk = "mic",
  remove_NAs = TRUE,
  gene_determinants = NULL,
  colour_by = "pheno",
  plot_breakpoints = FALSE,
  organism = NULL,
  break_guide = "EUCAST 2024",
  break_type = "ECOFF",
  drug = NULL,
  sort_intersections_by = "degree",
  sort_intersections = "ascending",
  show_intersect_size = TRUE,
  intersect_counts = TRUE,
  heights = NULL,
  colour_values = c(S = "#66c2a5", I = "#fdae61", R = "#d53e4f")
)
}
\arguments{
\item{binary_matrix}{A binary matrix where rows represent samples and columns represent genetic determinants or phenotypic resistance data.}

\item{min_set_size}{The minimum size of a set to be included in the plot. Default is 10.}

\item{mic_disk}{A character string specifying whether to use 'mic' or 'disk' for the y-axis. Must be either \code{"mic"} or \code{"disk"}. Default is \code{"mic"}.}

\item{remove_NAs}{Logical. If \code{TRUE}, removes rows with missing values in the selected \code{mic_disk} column. Default is \code{TRUE}.}

\item{gene_determinants}{A character vector specifying which genetic determinants to include. If \code{NULL}, all genes are used. Default is \code{NULL}.}

\item{colour_by}{A character string specifying the column used for colour mapping in the plot. Default is \code{"pheno"}.}

\item{plot_breakpoints}{Logical. If \code{TRUE}, overlays MIC or disk diffusion breakpoints on the plot. Default is \code{FALSE}.}

\item{organism}{A character string specifying the organism, used when plotting breakpoints.}

\item{break_guide}{A character string specifying the breakpoint guideline (e.g., \code{"EUCAST 2024"}). Default is \code{"EUCAST 2024"}.}

\item{break_type}{A character string specifying the breakpoint type (e.g., \code{"ECOFF"}). Default is \code{"ECOFF"}.}

\item{drug}{A character string specifying the antimicrobial agent to be analysed.}

\item{colour_values}{A named vector specifying colours for different resistance categories (\code{S}, \code{I}, \code{R}). Default is \code{c(S="#66c2a5", I="#fdae61", R="#d53e4f")}.}
}
\value{
A \code{ggplot} object displaying the Upset plot.
}
\description{
This function generates an Upset plot using a binary matrix of genetic determinants and antimicrobial resistance (AMR) data. Optionally, MIC or disk diffusion breakpoints can be overlaid on the plot.
}
\examples{
# Example usage:

ecoli_geno <- import_amrfp(ecoli_geno_raw, "Name")

binary_matrix<- get_binary_matrix(geno_table=ecoli_geno, 
              pheno_table=ecoli_ast, 
              antibiotic="Ciprofloxacin", 
              drug_class_list=c("Quinolones"), 
              sir_col="pheno", 
              keep_assay_values=TRUE, 
              keep_assay_values_from = "mic"
           )

amr_complexUpset(binary_matrix)

}
