% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amrfp_drugs_table}
\alias{amrfp_drugs_table}
\title{NCBI Subclass mapping to drug class}
\format{
\code{amrfp_drugs_table} A data frame with 21 rows and 2 columns.

Columns include:
\itemize{
\item \code{AMRFP_Subclass}: NCBI term
\item \code{drug_class}: Name of drug class as it should appear in imported genotype table.
}
}
\usage{
amrfp_drugs_table
}
\description{
Mapping of NCBI refgene / AMRfinderplus Subclass terms that are not present in the AMR package as drug class terms. Used internally when importing AMRfinderplus results into AMRgen genotype table format.
}
\keyword{datasets}
