% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.gene.R
\name{as.gene}
\alias{as.gene}
\title{Gene Class and AMR Parsing Functions}
\usage{
as.gene(x)
}
\arguments{
\item{x}{A character vector to be converted to a "gene" class.}
}
\value{
For \code{as.gene}, an object of class "gene". For \code{\link[=import_amrfp]{import_amrfp()}}, a data frame with parsed AMR data.
}
\description{
Functions to create a custom "gene" class and parse AMR data.
}
\examples{
\dontrun{
# Create a gene object
gene <- as.gene(c("gene1", "gene2"))
print(gene)

# Parse AMR data
parsed_data <- import_amrfp("path/to/input_table.tsv", "SampleID")
}
}
