% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ast_CLI_public}
\alias{ast_CLI_public}
\title{Clindamycin MIC data for 5914 \emph{Staphylococcus aureus} isolates}
\format{
\code{ast_CLI_public} A data frame with 5914 rows and 34 columns:
\itemize{
\item \code{id}: Sample identifier, imported from the \verb{#BioSample} column in the raw input.
\item \code{drug_agent}: Antibiotic code, interpreted from \code{Antibiotic} using \code{as.ab}, used to interpret \code{ecoff} and \code{pheno} columns.
\item \code{mic}: Minimum inhibitory concentration, formatted using \code{as.mic}, used to interpret \code{ecoff} and \code{pheno} columns.
\item \code{disk}: Disk diffusion zone, formatted using \code{as.disk}, used to interpret \code{ecoff} and \code{pheno} columns.
\item \code{pheno_eucast}: S/I/R classification according to EUCAST, interpreted using \code{as.sir}.
\item \code{ecoff}: WT/NWT classification, interpreted using \code{as.sir}.
\item \code{guideline}: Interpretation guidelines used to interpret \code{ecoff} and \code{pheno} columns.
\item \code{method}: Test method, one of: "E-test', "Unknown', "agar dilution', "broth dilution', "disk diffusion'.
\item \code{platform}: Testing platform, one of "BD Phoenix', "Microscan', "Phoenix', "Sensititre', "Vitek'.
\item \code{pheno_provided}: S/I/R interpretation as provided in the raw input.
\item \code{spp_pheno}: Species identifier, interpreted from \verb{Scientific name} using \code{as.mo}, used to interpret \code{ecoff} and \code{pheno} columns.
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pathogens/ast}

\url{https://www.ebi.ac.uk/amr}
}
\usage{
ast_CLI_public
}
\description{
Used to investigate genetic determinants of clindamycin resistance.
Downloaded from NCBI and EBI.
}
\keyword{datasets}
