% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_pheno}
\alias{btESBL_pheno}
\title{blantyreESBL AST data}
\format{
\code{btESBL_pheno} A data frame with 609 rows and 9 columns:
\itemize{
\item \code{...1}:  Row count.
\item \code{supplier_name}:  Strain ID.
\item \code{organism}:  organism.
\item \code{amikacin}:  amikacin antimicrobial susceptibility phenotype (S/R)
\item \code{chloramphenicol}:  chloramphenicol antimicrobial susceptibility phenotype (S/R)
\item \code{ciprofloxacin}:  ciprofloxacin antimicrobial susceptibility phenotype (S/R)
\item \code{cotrimoxazole}:  cotrimoxazole antimicrobial susceptibility phenotype (S/R)
\item \code{gentamicin}:  gentamicin antimicrobial susceptibility phenotype (S/R)
\item \code{meropenem}:  meropenem antimicrobial susceptibility phenotype (S/R)
}
}
\source{
\url{https://github.com/joelewis101/blantyreESBL/raw/refs/heads/main/data/btESBL_pheno.rda}
}
\usage{
btESBL_pheno
}
\description{
The antimicrobial susceptibility testing (AST) data from the blantyreESBL Github from Dr. Joseph Lewis.
Antimicrobial sensitivity testing (AST) was carried out on a subset of isolates using the disc-diffusion method
using British Society for Antimicrobial Chemotherapy (BSAC) guidelines (https://bsac.org.uk/).
AST was carried out for meropenem, amikacin, chloramphenicol, ciprofloxacin, co-trimoxazole and gentamicin.
However this dataset only contains the raw phenotype data (S/R).
}
\keyword{datasets}
