% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{checkBreakpoints}
\alias{checkBreakpoints}
\title{Check and Retrieve Breakpoints for an Antibiotic}
\usage{
checkBreakpoints(
  species,
  guide = "EUCAST 2024",
  antibiotic,
  assay = "MIC",
  bp_site = NULL
)
}
\arguments{
\item{species}{A character string representing the species of interest (e.g., "Escherichia coli").}

\item{guide}{A character string indicating the guideline for breakpoints (default, "EUCAST 2024").}

\item{antibiotic}{A character string indicating the antibiotic for which to check breakpoints (e.g., "Ciprofloxacin").}

\item{assay}{A character string specifying the assay type (either "MIC" or "Disk"). Default is "MIC".}

\item{bp_site}{A character string specifying the breakpoint site to use (optional). If provided, the function uses this site; otherwise, if different breakpoints are specified for different sites it selects the one with the highest susceptibility breakpoint.}
}
\value{
A list containing:
\itemize{
\item \code{breakpoint_S}: The susceptibility breakpoint (e.g., MIC or disk size).
\item \code{breakpoint_R}: The resistance breakpoint (e.g., MIC or disk size).
\item \code{bp_standard}: The breakpoint site used, if multiple breakpoints are set in the guidelines.
}
}
\description{
This function checks the clinical breakpoints for a specified antibiotic and species using the \code{getBreakpoints} function.
It handles cases where multiple breakpoint sites exist and uses the specified site or the one with the highest susceptibility
breakpoint.
}
\details{
The function first attempts to retrieve breakpoints using the \code{getBreakpoints} function. If multiple breakpoint sites are found, it handles the situation by:
\itemize{
\item Using the specified site if it exists.
\item Selecting the breakpoint with the highest susceptibility value if the specified site is not found.
\item Returning a message about the selected site and breakpoint values.
}
}
\examples{
checkBreakpoints(
  species = "Escherichia coli", guide = "EUCAST 2024",
  antibiotic = "Ciprofloxacin", assay = "MIC"
)

}
