% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compare_geno_pheno_id}
\alias{compare_geno_pheno_id}
\title{Compare Genotype and Phenotype Data by Sample ID}
\usage{
compare_geno_pheno_id(
  geno_data,
  pheno_data,
  geno_sample_col = NULL,
  pheno_sample_col = NULL,
  rename_id_cols = F
)
}
\arguments{
\item{geno_data}{A data frame containing the genotype data. The first column (or the column specified by
\code{geno_sample_col}) should represent sample IDs.}

\item{pheno_data}{A data frame containing the phenotype data. The first column (or the column specified by
\code{pheno_sample_col}) should represent sample IDs.}

\item{geno_sample_col}{A string specifying the name of the column in \code{geno_data} that contains the sample IDs.
Defaults to the first column of \code{geno_data}.}

\item{pheno_sample_col}{A string specifying the name of the column in \code{pheno_data} that contains the sample IDs.
Defaults to the first column of \code{pheno_data}.}

\item{rename_id_cols}{A logical value indicating whether to rename the sample ID columns in both \code{geno_data} and
\code{pheno_data} to "id". Defaults to \code{FALSE}.}
}
\value{
A list with the following elements:
\item{pheno_unique}{A vector of sample IDs that are unique to the phenotype dataset.}
\item{geno_unique}{A vector of sample IDs that are unique to the genotype dataset.}
\item{overlap_ids}{A vector of sample IDs that are common to both datasets.}
\item{geno_matched}{A data frame of the genotype data filtered to only include the samples with matching IDs.}
\item{pheno_matched}{A data frame of the phenotype data filtered to only include the samples with matching IDs.}
}
\description{
This function compares genotype (\code{geno_data}) and phenotype (\code{pheno_data}) datasets based on their sample IDs.
It identifies sample IDs that are unique to either dataset, as well as those that overlap. Optionally,
it renames the sample ID columns in both datasets for consistency.
}
\examples{
# Example usage
geno_table <- import_amrfp(ecoli_geno_raw, "Name")

# example phenotype data
head(ecoli_ast)

result <- compare_geno_pheno_id(geno_table,
                                ecoli_ast,
                                geno_sample_col = "Name",
                                pheno_sample_col = "id")
print(result$pheno_unique)
print(result$geno_unique)
print(result$overlap_ids)
print(result$geno_matched)
print(result$pheno_matched)

}
