% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{concordance_from_tables}
\alias{concordance_from_tables}
\title{Assess concordance between tables of observed and predicted phenotypes}
\usage{
concordance_from_tables(
  pheno_table,
  pheno_pred_table,
  true_SIR_col = "pheno_eucast",
  true_ecoff_col = "ecoff",
  pred_SIR = "clinical category",
  pred_ecoff = "phenotype",
  sample_col = "id",
  drug_col = "drug_agent",
  measure_col = "mic"
)
}
\arguments{
\item{pheno_table}{A tibble or data frame containing real (observed) phenotype data, in the format output by \code{\link[=import_pheno]{import_pheno()}}.}

\item{pheno_pred_table}{A tibble or data frame containing predicted phenotype calls, e.g. AMRrules predictions imported in AMRgen using \code{\link[=import_amrrules_predictions]{import_amrrules_predictions()}}.}

\item{true_SIR_col}{Character. Name of the column containing S/I/R calls interpreted from real data against clinical breakpoints (default \code{"pheno_eucast"}).}

\item{true_ecoff_col}{Character. Name of the column containing WT/NWT calls interpreted from real data against ECOFF (default \code{"ecoff"}).}

\item{pred_SIR}{Character. Name of the column containing S/I/R calls predicted from genotypes (default \code{"clinical category"}).}

\item{pred_ecoff}{Character. Name of the column containing WT/NWT calls predicted from genotypes (default \code{"phenotype"}).}

\item{sample_col}{Character. Name of the column containing sample identifiers. This must be the same in both tables (default \code{"id"}).}

\item{drug_col}{Character. Name of the column containing drug agent identifiers. This must be the same in both tables (default \code{"drug_agent"}).}

\item{measure_col}{Character. Name of the column containing observed MIC or disk measurements for plotting. Valid options are \code{"mic"} or \code{"disk"} (default \code{"mic"}).}
}
\value{
A named list with the following elements:
\describe{
\item{obs_pred}{A copy of the pheno_table with the predictions from pheno_pred_table merged in.}
\item{metrics}{A tibble listing concordance metrics, comparing observed vs predicted R and/or NWT calls, for all drugs with matched samples in both input tables.}
\item{drugs_R}{A long form tibble listing the number of samples with each combination of observed and predicted binary calls for R, for all drugs with matched samples in both input tables.}
\item{drugs_NWT}{A long form tibble listing the number of samples with each combination of observed and predicted binary calls for NWT, for all drugs with matched samples in both input tables.}
\item{plot_R}{A list containing two-panel plots summarising observed vs predicted clinical categories, one for each drug.}
\item{plot_NWT}{A list containing two-panel plots summarising observed vs predicted WT/NWT categories, one for each drug.}
\item{plot_R_dist}{A list containing plots of the assay value distributions (MIC or disk zones), coloured by S/I/R prediction, one for each drug.}
\item{plot_NWT_dist}{A list containing plots of the assay value distributions (MIC or disk zones), coloured by WT/NWT prediction, one for each drug.}
}
}
\description{
Calculates concordance from two data frames in standard phenotype table format, one with observed phenotype data and the other with predicted phenotype calls (e.g. output from AMRrules).
}
\examples{
\dontrun{
concordance_from_tables(true_phenotypes, predicted_phenotypes)
}
}
