% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_amrfp.R
\name{convert_mutation}
\alias{convert_mutation}
\title{Convert mutation string based on method}
\usage{
convert_mutation(gene_symbol_col, method_col)
}
\arguments{
\item{gene_symbol_col}{A character vector representing the 'Gene symbol'
column (the mutation string).}

\item{method_col}{A character vector representing the 'Method' column.}
}
\value{
A character vector containing the formatted mutation strings
(e.g., "Ala123Trp") or NA if not applicable/match.
}
\description{
This function takes a mutation string (e.g., "gene_REF123ALT") and a
mutation method, then extracts and converts parts of the mutation string.
Specifically designed for use within \code{dplyr::mutate()}.
}
