% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cro_comparison}
\alias{cro_comparison}
\title{N. gonorrhoeae Euro-GASP ceftriaxone data vs EUCAST reference distribution}
\format{
\code{cro_comparison} A data frame with 14 rows and 3 columns:
\itemize{
\item \code{value}: MIC value
\item \code{user}: Number of samples with each MIC value, from Euro-GASP data (total n=5,358 isolates)
\item \code{eucast}: Number of samples with each MIC value, from EUCAST reference data (total n=47,903 isolates)
}
}
\source{
https://mic.eucast.org/
}
\usage{
cro_comparison
}
\description{
Minimum inhibitory concentration (MIC) distributions for ceftriaxone in
\emph{Neisseria gonorrhoeae} isolates. EUCAST reference distribution (retrieved
March 2026 using \code{\link[=compare_mic_with_eucast]{compare_mic_with_eucast()}}) compared with data from three
European Gonococcal Antimicrobial
Surveillance Programme (Euro-GASP) genomic surveys (2013, 2018, 2020), see
Harris \emph{et al.} (2018) \url{https://doi.org/10.1016/S1473-3099(18)30225-1},
Sánchez-Busó \emph{et al.} (2022) \url{https://doi.org/10.1016/S2666-5247(22)00044-1}, and
Golparian \emph{et al.} (2024) \url{https://doi.org/10.1016/S2666-5247(23)00370-1}.
}
\keyword{datasets}
