% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DASSIM_geno}
\alias{DASSIM_geno}
\title{DASSIM genotype data (AMRfinderplus)}
\format{
\code{DASSIM_geno} A data frame with 12414 rows and 31 columns:
\itemize{
\item \code{Name}:  Name.
\item \code{gene}:  gene.
\item \code{mutation}:  mutation.
\item \code{node}:  node.
\item \code{marker}:  marker.
\item \code{marker.label}:  marker.label.
\item \code{drug_agent}:  drug_agent.
\item \code{drug_class}:  drug_class.
\item \verb{Protein id}:  Protein id.
\item \verb{Contig id}:  Contig id.
\item \code{Start}:  Start.
\item \code{Stop}:  Stop.
\item \code{Strand}:  Strand.
\item \verb{Gene symbol}:  Gene symbol.
\item \verb{Element name}:  Element name.
\item \code{Scope}:  Scope.
\item \code{Type}:  Type.
\item \code{Subtype}:  Subtype.
\item \code{Class}:  Class.
\item \code{Subclass}:  Subclass.
\item \code{Method}:  Method.
\item \verb{Target length}:  Target length.
\item \verb{Reference sequence length}:  Reference sequence length.
\item \verb{\% Coverage of reference}:  \% Coverage of reference.
\item \verb{\% Identity to reference}:  \% Identity to reference.
\item \verb{Alignment length}:  Alignment length.
\item \verb{Closest reference accession}:  Closest reference accession.
\item \verb{Closest reference name}:  Closest reference name.
\item \verb{HMM accession}:  HMM accession.
\item \verb{HMM description}:  HMM description.
\item \verb{variation type}:  variation type.
}
}
\source{
\url{https://www.ebi.ac.uk/ena/browser/view/PRJEB26677}

\url{https://www.ebi.ac.uk/ena/browser/view/PRJEB36486}
}
\usage{
DASSIM_geno
}
\description{
The DASSIM dataset screened for Antimicrobial resistance genes (ARGs) using AMRfinderplus v4.0.23
}
\keyword{datasets}
