% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DASSIM_pheno_raw}
\alias{DASSIM_pheno_raw}
\title{DASSIM phenotype metadata}
\format{
\code{DASSIM_pheno_raw} A data frame with 840 rows and 15 columns:
\itemize{
\item \code{Strain_ID}:  Strain_ID.
\item \code{ST}:  Sequence Type.
\item \code{Species}:  Species.
\item \verb{CHL susceptibility}:  susceptibility to chloramphenicol
\item \code{CatA1}:  CatA1 presence/absence
\item \code{CatA2}:  CatA2 presence/absence
\item \code{CatB4}:  CatB4 presence/absence
\item \code{CatB3}:  CatB3 presence/absence
\item \code{FloR}:  FloR presence/absence
\item \code{CmlA1}:  CmlA1 presence/absence
\item \code{CmlA5}:  CmlA5 presence/absence
\item \code{study}:  study presence/absence
\item \verb{ENA accession}:  ENA accession.
\item \code{lane}:  lane.
\item \code{seq}:  seq.
}
}
\source{
\url{https://static-content.springer.com/esm/art\%3A10.1038\%2Fs41467-024-53391-2/MediaObjects/41467_2024_53391_MOESM4_ESM.xls}
}
\usage{
DASSIM_pheno_raw
}
\description{
Additional metadata relating to the DASSIM isolates found in Supplementary Data 1 of the following paper:
Molecular mechanisms of re-emerging chloramphenicol susceptibility in extended-spectrum beta-lactamase-producing
Enterobacterales. Nat Commun 15, 9019 (2024). Fabrice E Graf et al.
}
\keyword{datasets}
