% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ebi.R
\name{download_ebi_geno}
\alias{download_ebi_geno}
\title{Download antimicrobial genotype data from the EBI AMR Portal}
\usage{
download_ebi_geno(
  user_genus = NULL,
  user_release = NULL,
  user_antibiotic_name = NULL
)
}
\arguments{
\item{user_genus}{String specifying a bacterial genus to download data for (default NULL, will pull all taxa)}

\item{user_release}{String specifying the data release to download (default NULL, will pull latest release)}

\item{user_antibiotic_name}{String specifying an antibiotic to download data for (default NULL, will pull all antibiotics).}
}
\value{
A data frame containing EBI genotype data
}
\description{
This function will retrieve genotype data from the EBI AMR Portal, https://www.ebi.ac.uk/amr. The portal uses AMRfinderplus to identify AMR-associated genotypes, but the results are processed and not all fields returned by AMRfinderplus are included. See https://www.ebi.ac.uk/amr/about/#AMR-Genotypes for more information.
}
\examples{
\dontrun{
amp_sal_ebi <- import_ebi(
    user_genus="Salmonella",
    user_release="2025-12",
    user_antibiotic_name="ampicillin"
)
}
}
\keyword{internal}
