% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli_ast}
\alias{ecoli_ast}
\title{E. coli NCBI AST example data, re-interpreted with AMR package}
\format{
\subsection{\code{ecoli_ast}}{

A data frame with 15320 rows and 24 columns representing reinterpreted data from NCBI AST Browser:
\describe{
\item{\code{id}}{sample identifier, imported from BioSample column in raw input}
\item{spp_pheno}{species identifier, interpreted from \verb{Scientific name} using as.mo, used to interpret ecoff and pheno columns}
\item{drug_agent}{antibiotic code, interpreted from \code{Antibiotic} using as.ab, used to interpret ecoff and pheno columns}
\item{ecoff}{WT/NWT classification, interpreted using as.sir}
\item{pheno}{S/I/R classification, interpreted using as.sir}
\item{mic}{minimum inhibitory concentration, formated using as.mic, used to interpret ecoff and pheno columns}
\item{disk}{disk diffusion zone, formated using as.disk, used to interpret ecoff and pheno columns}
\item{guideline}{interpretation guidelines used to interpret ecoff and pheno columns}
\item{\verb{Scientific name}}{species identifier, from input file}
\item{Antibiotic}{antibiotic name, from input file}
\item{\verb{Testing standard}}{interpretation standard (EUCAST or CLSI), from input file}
\item{\verb{Measurement sign}}{measurement sign (>, <, =, etc) relating to MIC measurement, from input file}
\item{\code{MIC (mg/L)}}{minimum inhibitory concentration, from input file}
\item{\verb{Disk diffusion (mm)}}{disk diffusion zone, from input file}
\item{\verb{Resistance phenotype}}{resistance call (SIR), from input file}

...
}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pathogens/ast}
}
\usage{
ecoli_ast
}
\description{
A subset of E. coli phenotype data from the NCBI AST browser
}
\keyword{datasets}
