% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli_ast_raw}
\alias{ecoli_ast_raw}
\title{E. coli NCBI AST example data}
\format{
\subsection{\code{ecoli_ast_raw}}{

A data frame with 10 rows and 17 columns representing unprocessed data from NCBI AST Browser:
\describe{
\item{\verb{#BioSample}}{sample identifier}
\item{\verb{Scientific name}}{species identifier}
\item{Antibiotic}{antibiotic name}
\item{\verb{Testing standard}}{interpretation standard (EUCAST or CLSI)}
\item{\verb{Measurement sign}}{measurement sign (>, <, =, etc) relating to MIC measurement}
\item{\code{MIC (mg/L)}}{minimum inhibitory concentration}
\item{\verb{Disk diffusion (mm)}}{disk diffusion zone}
\item{\verb{Resistance phenotype}}{resistance call (SIR) as submitted}

...
}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pathogens/ast}
}
\usage{
ecoli_ast_raw
}
\description{
A subset of E. coli phenotype data from the NCBI AST browser
}
\keyword{datasets}
