% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli_cip_mic_data}
\alias{ecoli_cip_mic_data}
\title{EUCAST Reference distribution for Ciprofloxacin in E. coli}
\format{
A data frame with 19 rows and 2 columns. It provides MIC distributions
from EUCAST in the form of counts per value.

Columns include:
\itemize{
\item \code{mic}: MIC value.
\item \code{count}: Count of samples with this MIC value, downloaded from EUCAST (Feb 2026).
}
}
\source{
\url{https://mic.eucast.org/}
}
\usage{
ecoli_cip_mic_data
}
\description{
Data frame containing EUCAST reference distribution for ciprofloxacin in E. coli, downloaded using \link{get_eucast_mic_distribution}.
}
\keyword{datasets}
