% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli_cip_vs_ref}
\alias{ecoli_cip_vs_ref}
\title{E. coli Ciprofloxacin MIC Distribution Example Data}
\format{
An object of class \code{compare_eucast} with 32 rows and
3 columns. It provides MIC distributions
from EUCAST and public AST data extracted from \link{ecoli_ast}
in the form of counts per value.

Columns include:
\itemize{
\item \code{value}: MIC value.
\item \code{user}: Count of samples with this MIC value, from the example data \link{ecoli_ast}.
\item \code{eucast}: Count of samples with this MIC value, downloaded from EUCAST (Feb 2026).
}
}
\source{
\url{https://mic.eucast.org/}
}
\usage{
ecoli_cip_vs_ref
}
\description{
Ciprofloxacin MIC distributions for E. coli, calculated from public data
and compared with the EUCAST reference distribution.
}
\keyword{datasets}
