% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli_geno_raw}
\alias{ecoli_geno_raw}
\title{E. coli genotype example data}
\format{
\subsection{\code{ecoli_geno_raw}}{

A data frame with 53591 rows and 10 columns representing genotyping results from AMRfinderplus:
\describe{
\item{Name}{sample identifier}
\item{\verb{Gene symbol}}{gene symbol in NCBI refgene}
\item{\verb{Hierarchy node}}{node in NCBI hierarchy}
\item{Class, Subclass}{drug class/es associated with marker (in NCBI refgene)}
\item{\verb{\% Coverage of reference sequence},\% Identity to reference sequence\verb{, }Accession of closest sequence` }{sequence match information}

...
}
}
}
\source{
\url{https://github.com/ncbi/amr/wiki/Interpreting-results}
}
\usage{
ecoli_geno_raw
}
\description{
Genotypes called using AMRfinderplus (v3.12.8, DB 2024-01-31.1), sourced from AllTheBacteria project
}
\keyword{datasets}
