% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eucast_supported_ab_distributions.R
\name{eucast_supported_ab_distributions}
\alias{eucast_supported_ab_distributions}
\title{Retrieve Available Antimicrobial Wild Type Distributions from EUCAST}
\usage{
eucast_supported_ab_distributions(...)
}
\arguments{
\item{...}{arguments passed on to function, currently unused}
}
\description{
Run this function to get an updated list of antimicrobial distributions currently supported by EUCAST. This retrieves live info from \url{https://mic.eucast.org}.
}
\examples{
eucast_supported_ab_distributions()
}
