% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eurogasp_pheno_raw}
\alias{eurogasp_pheno_raw}
\title{N. gonorrhoeae Euro-GASP Phenotype Data Use Case 1}
\format{
\code{eurogasp_pheno_raw} A data frame with 5,361 rows and 5 columns:
\itemize{
\item \code{id}: Sample identifier (ENA run accession).
\item \code{Azithromycin}: MIC value in mg/L for azithromycin (numeric; n = 5,055).
\item \code{Ciprofloxacin}: MIC value in mg/L for ciprofloxacin (numeric; n = 5,360).
\item \code{Cefixime}: MIC value in mg/L for cefixime (character, to preserve inequality prefixes such as \verb{<0.016}).
\item \code{Ceftriaxone}: MIC value in mg/L for ceftriaxone (numeric; n = 5,361).
}
}
\source{
ENA projects \href{https://www.ebi.ac.uk/ena/browser/view/PRJEB9227}{PRJEB9227},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJEB34068}{PRJEB34068},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJEB58139}{PRJEB58139}.
}
\usage{
eurogasp_pheno_raw
}
\description{
Minimum inhibitory concentration (MIC) data for \emph{Neisseria gonorrhoeae} isolates
from three European Gonococcal Antimicrobial Surveillance Programme (Euro-GASP)
genomic surveys (2013, 2018, 2020), from
Harris \emph{et al.} (2018) \url{https://doi.org/10.1016/S1473-3099(18)30225-1},
Sánchez-Busó \emph{et al.} (2022) \url{https://doi.org/10.1016/S2666-5247(22)00044-1}, and
Golparian \emph{et al.} (2024) \url{https://doi.org/10.1016/S2666-5247(23)00370-1}.
}
\keyword{datasets}
