% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ncbi.R
\name{export_ncbi_biosample}
\alias{export_ncbi_biosample}
\title{Import/Export BioSample Antibiograms}
\usage{
export_ncbi_biosample(data, file, overwrite = FALSE)
}
\arguments{
\item{data}{data set containing SIR results}

\item{file}{file path}

\item{overwrite}{\link{logical} to overwrite existing file}
}
\description{
Output phenotype data to \href{https://www.ncbi.nlm.nih.gov/biosample/docs/antibiogram/}{NCBI BioSample Antibiograms}.
}
