% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ncbi.R
\name{format_ebi_json}
\alias{format_ebi_json}
\title{Generate EBI antibiogram submission in JSON}
\usage{
format_ebi_json(
  ebi_antibiogram_table,
  breakpoint_version,
  submission_account,
  domain = "self.ExampleDomain",
  output_dir
)
}
\arguments{
\item{ebi_antibiogram_table}{A data frame in the format output by
\code{\link[=export_ebi_ast]{export_ebi_ast()}}.}

\item{breakpoint_version}{Character string specifying the
breakpoint version used for interpretation (e.g. \code{"EUCAST 2024"}).}

\item{submission_account}{Character string specifying the Webin
submission account identifier (e.g. \code{"Webin-###"}).}

\item{domain}{Character string specifying the domain used in the
submission metadata (default \code{"self.ExampleDomain"}).}

\item{output_dir}{Character string specifying the directory where JSON
files should be written.}
}
\value{
Invisibly returns \code{NULL}. The function prints JSON-formatted
AMR submission records to file.
}
\description{
Converts the tabular output of \code{\link[=export_ebi_ast]{export_ebi_ast()}} into JSON
files formatted for submission to EBI as BioSample data
(https://www.ebi.ac.uk/amr/amr_submission_guide/). Each row of the input
dataset is converted into JSON records and printed to file.
}
\details{
The function iterates over each biosample in \code{ebi_antibiogram_table} and
constructs
a nested JSON object describing the antimicrobial susceptibility
testing result. Each record contains antibiotic metadata, AST
standards, measurement values, and resistance phenotype information.

JSON formatting is performed using \code{jsonlite::toJSON()} with
\code{pretty = TRUE} and \code{auto_unbox = TRUE} .
}
\examples{
\dontrun{
format_ebi_json(
  ast_dataset,
  breakpoint_version = "EUCAST 2015",
  submission_account = "Webin-###",
  domain = "self.ExampleDomain",
  output_dir = "/path/to/output/"
)
}

}
