% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pheno.R
\name{format_pheno}
\alias{format_pheno}
\title{Import and process antimicrobial phenotype data from a generic format}
\usage{
format_pheno(
  input,
  sample_col = "id",
  species = NULL,
  species_col = "spp_pheno",
  ab = NULL,
  ab_col = "drug",
  mic_col = "mic",
  disk_col = "disk",
  pheno_cols = c("ecoff", "pheno_eucast", "pheno_clsi", "pheno_provided"),
  method_col = "method",
  platform_col = "platform",
  source_col = "source",
  guideline_col = "guideline",
  interpret_eucast = FALSE,
  interpret_clsi = FALSE,
  interpret_ecoff = FALSE,
  rename_cols = TRUE
)
}
\arguments{
\item{input}{A string representing a dataframe, or a path to an input file, containing the AST data in long-form antibiogram format (one row per sample and test). This might be a file containing the content of an EBI/NCBI AST dataset previously processed using \code{\link[=import_ebi_pheno]{import_ebi_pheno()}}, \code{\link[=import_ncbi_pheno]{import_ncbi_pheno()}}, or \code{\link[=import_pheno]{import_pheno()}} functions, or files with a similar format/structure but with different column names.}

\item{sample_col}{(optional, default \code{"id"}) String indicating the name of the input data column that provides the sample name. If the \code{rename_cols} parameter is set to \code{TRUE}, this column will be renamed as \code{id}.}

\item{species}{(optional, default \code{NULL}) String indicating the name of the single species to which all samples belong. Use this if you want to interpret assay measurements but the input file does not contain a column indicating the species for each sample (called \code{species} or a name specified by the \code{species_col} parameter).}

\item{species_col}{(optional, default \code{"species"}) String indicating the name of the input data column that provides a species name. If provided, this column will be converted to micro-organism class \code{mo} via \code{\link[AMR:as.mo]{AMR::as.mo()}}. If the \code{rename_cols} parameter is set to \code{TRUE}, this column will also be renamed as \code{spp_pheno}. If interpretation is switched on, this column will be used to identify the appropriate  breakpoints for interpretation of each row in the data table.}

\item{ab}{(optional, default \code{NULL}) String indicating the name of a single antibiotic to use for phenotype interpretation. Use this if you want to interpret assay measurements but the input file does not contain a column indicating the drug for each sample (called \code{drug} or a name specified by the \code{ab_col} parameter).}

\item{ab_col}{(optional, default \code{"drug"}) String indicating the name of the input data column that provides a drug name. If provided, this column will be converted to antibiotic class \code{ab} via \code{\link[AMR:as.ab]{AMR::as.ab()}}. If the \code{rename_cols} parameter is set to \code{TRUE}, this column will also be renamed as \code{drug}. If interpretation is switched on, this column will be used to identify the appropriate breakpoints for interpretation of each row in the data table.}

\item{mic_col}{(optional, default \code{"mic"}) String indicating the name of the input data column that provides MIC measurements. If provided, this column will be converted to MIC class \code{mic} via \code{\link[AMR:as.mic]{AMR::as.mic()}}. If the \code{rename_cols} parameter is set to \code{TRUE}, this column will also be renamed as \code{mic}. If interpretation is switched on, the MIC values will be interpreted against clinical breakpoints.}

\item{disk_col}{(optional, default \code{"disk"}) String indicating the name of the input data column that provides disk diffusion zone measurements. If provided, this column will be converted to disk diffusion class \code{disk} via \code{\link[AMR:as.disk]{AMR::as.disk()}}. If the \code{rename_cols} parameter is set to \code{TRUE}, this column will also be renamed as \code{disk}. If interpretation is switched on, the zone values will be interpreted against clinical breakpoints.}

\item{pheno_cols}{(optional, default \code{c("ecoff", "pheno_eucast", "pheno_clsi", "pheno_provided")}) String, or vector of strings, indicating the name/s of the input data column/s that provide interpreted phenotypes (with column values encoded as \code{S/I/R} or \code{WT/NWT}). If provided, these columns will be converted to class \code{sir} via \code{\link[AMR:as.sir]{AMR::as.sir()}}.}

\item{method_col}{(optional, default \code{"method"}) String indicating the name of the input data column that indicates the testing method used (e.g. \code{"broth dilution"}, \code{"disk diffusion"}). If the \code{rename_cols} parameter is set to \code{TRUE}, this column will also be renamed as \code{method}.}

\item{platform_col}{(optional, default \code{"platform"}) String indicating the name of the input data column that indicates the testing platform used (e.g. \code{"Vitek"}, \code{"Sensititre"}). If the \code{rename_cols} parameter is set to \code{TRUE}, this column will also be renamed as \code{platform}.}

\item{source_col}{(optional, default \code{"source"}) String indicating the name of the input data column that indicates the source of the dataset (e.g. \code{"BioProject"}, \code{"PMID"}). If the \code{rename_cols} parameter is set to \code{TRUE}, this column will also be renamed as \code{source}.}

\item{guideline_col}{(optional, default \code{"guideline"}) String indicating the name of the input data column that indicates the guideline used for testing (e.g. \code{"EUCAST"}, \code{"CLSI"}). If the \code{rename_cols} parameter is set to \code{TRUE}, this column will also be renamed as \code{guideline}.}

\item{interpret_eucast}{A logical value (default is \code{FALSE}). If \code{TRUE}, the function will interpret the susceptibility phenotype (SIR) for each row based on the MIC or disk diffusion values, against EUCAST human breakpoints. These will be reported in a new column \code{pheno_eucast}, of class \code{sir}.}

\item{interpret_clsi}{A logical value (default is \code{FALSE}). If \code{TRUE}, the function will interpret the susceptibility phenotype (SIR) for each row based on the MIC or disk diffusion values, against CLSI human breakpoints. These will be reported in a new column \code{pheno_clsi}, of class 'sir'.}

\item{interpret_ecoff}{A logical value (default is \code{FALSE}). If \code{TRUE}, the function will interpret the wildtype vs nonwildtype status for each row based on the MIC or disk diffusion values, against epidemiological cut-off (ECOFF) values. These will be reported in a new column \code{ecoff}, of class \code{sir} and coded as \code{NWT} (nonwildtype) or \code{WT} (wildtype).}

\item{rename_cols}{A logical value (default is \code{TRUE}). If \code{TRUE}, the function will rename the provided columns (specified by \code{sample_col}, \code{ab_col}, \code{mic_col}, \code{disk_col}, \code{species_col}) to the default names expected by AMRgen functions (\code{id}, \code{drug}, \code{mic}, \code{disk}, \code{spp_pheno}), to match those output by the other \code{\link[=import_pheno]{import_pheno()}} functions.}
}
\value{
A data frame with the processed AST data, including additional columns:
}
\description{
This function attempts to import antibiotic susceptibility testing (AST) data in long-form antibiogram format (one row per sample and test), suitable for downstream use with AMRgen analysis functions. It assumes that the input file is a tab-delimited text file (e.g., TSV) or CSV (which may be compressed) and parses relevant columns (antibiotic names, species names, MIC or disk data, S/I/R calls) into suitable classes using the AMR package. It optionally can use the AMR package to interpret susceptibility phenotype (SIR) based on EUCAST or CLSI guidelines (human breakpoints and/or ECOFF).
}
\examples{
\dontrun{
# import and process AST data from EBI, write formatted data to file for later use
pheno <- import_ebi_pheno("EBI_AMR_data.csv.gz")
write_tsv(pheno,
  file = "EBI_AMR_data_processed.tsv.gz",
  interpret_eucast = TRUE, interpret_ecoff = TRUE
)

# read stored data and format the columns to the correct classes
pheno <- format_pheno("EBI_AMR_data_processed.tsv.gz")

# read in unprocessed E. coli AST data from non-standard format and interpret
pheno <- format_pheno("AMR_data.tsv",
  sample_col = "STRAIN", species = "E. coli",
  ab_col = "Antibiotic", mic_col = "MIC (mg/L)",
  interpret_eucast = TRUE, interpret_ecoff = TRUE
)
}
}
