% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{getBreakpoints}
\alias{getBreakpoints}
\title{Get Clinical Breakpoints for an Antibiotic}
\usage{
getBreakpoints(
  species,
  guide = "EUCAST 2024",
  antibiotic,
  type_filter = "human"
)
}
\arguments{
\item{species}{A character string representing the species of interest (e.g., "Escherichia coli").}

\item{guide}{A character string indicating the guideline for breakpoints (default, "EUCAST 2024").}

\item{antibiotic}{A character string indicating the antibiotic for which to retrieve breakpoints (e.g., "Ciprofloxacin").}

\item{type_filter}{A character string indicating the type of breakpoints to retrieve (e.g., "human"). Default is "human" which returns human clinical breakpoints, change to "ECOFF" to get the epidemiological cutoff.}
}
\value{
A data frame containing the clinical breakpoints for the specified species, antibiotic, and guideline.
If no exact match is found, the function attempts to retrieve breakpoints at the genus, family, and order levels.
}
\description{
This function retrieves the clinical breakpoints for a given species, antibiotic, and guideline, from the AMR package.
It attempts to find the breakpoints at various taxonomic levels (species, genus, family, order) if no direct match is found.
}
\examples{
getBreakpoints("Escherichia coli", "EUCAST 2024", "Ciprofloxacin")

}
