% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtdb.R
\name{gtdb.mo}
\alias{gtdb.mo}
\title{Get Microorganism from GTDB Species Name}
\usage{
gtdb.mo(species)
}
\arguments{
\item{species}{name of species, coerced with \code{\link[=as.mo]{as.mo()}}}
}
\value{
microorganism
}
\description{
Parse a character vector containing species names from GTDB output to get a valid microorganism code.
}
\examples{
gtdb.mo("Escherichia_A coli_BC")
}
