% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hAMRonization_function_v3.R
\name{harmonize_data}
\alias{harmonize_data}
\title{hamronize_data}
\usage{
harmonize_data(
  user_software_name,
  user_software_version,
  user_database_version,
  user_input_filename
)
}
\arguments{
\item{user_software_name}{the analysis software used to screen genome data for AMR determinants (must be amrfinderplus, rgi, or resfinder)}

\item{user_software_version}{the version of the analysis software used to screen genome data}

\item{user_database_version}{the version of the database used}

\item{user_input_filename}{the name of the genotypic AMR data file}
}
\value{
A data frame containing 'harmonized' AMR genotype data
}
\description{
hamronize_data
}
\examples{
\dontrun{
harmonize_data(
  user_software_name = "amrfinderplus",
  user_software_version = "3.12.8",
  user_input_filename = "ATB_Achromobacter_AFP.tsv",
  user_database_version = "2024-01-31.1"
)
}
}
