% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_amrfp.R
\name{import_amrfp_ebi_ftp}
\alias{import_amrfp_ebi_ftp}
\title{Import EBI-processed AMRFinderPlus Genotypes from FTP}
\usage{
import_amrfp_ebi_ftp(input_table)
}
\arguments{
\item{input_table}{R object or file path for the input EBI genotype table (R object, or file path to a TSV or CSV file).}
}
\value{
A tibble containing the processed AMR elements, with harmonised gene names, drug agents, and drug classes. The output retains the original columns from the AMRFinderPlus table along with the newly mapped variables.
}
\description{
This function imports processed EBI-processed AMRFinderPlus genotyping results. The expected input is genotype data retrieved from the \href{https://ftp.ebi.ac.uk/pub/databases/amr_portal/releases/}{EBI AMR Portal FTP site} either directly or via the function \code{\link[=download_ebi]{download_ebi()}}.
Note that files downloaded from the \href{https://www.ebi.ac.uk/amr/data/?view=predictions}{EBI AMR Portal web browser} are formatted differently and can be imported using \link{import_amrfp_ebi_web}.
}
\details{
These data are pre-processed by EBI to match NCBI class/subclass to CARD's antibiotic resistance ontology (ARO), however for consistency this function will re-process the data to generate \code{drug_agent} and \code{drug_class} fields consistent with the \code{\link[=import_amrfp]{import_amrfp()}} function (the EBI fields \verb{antibiotic*} are also retained).
Note several AMRfinderplus fields are excluded from EBI files, including hierarchy node, method, percent identity and coverage; therefore unlike the \code{\link[=import_amrfp]{import_amrfp()}} function, this function cannot assign \verb{variation type} or \code{node}.

The function performs the following steps:
\itemize{
\item Reads the EBI-processed genotype table.
\item Maps AMRFinderPlus subclasses to standardised drug agent and drug class
names using \code{amrfp_drugs} (EBI-mappings are retained in \verb{antibiotic*} fields.)
\item Converts drug agent names to the \code{ab} class from the AMR package.
This processing ensures compatibility with downstream AMR analysis workflows.
}
}
\examples{
\dontrun{
# Download quinolone-related genotype data for E. coli, from EBI
ebi_geno_raw <- download_ebi(
  data = "genotype", species = "Escherichia coli",
  geno_subclass = "QUINOLONE"
)

# Format the file for import
ebi_geno <- import_amrfp_ebi_ftp(ebi_geno_raw)
}
}
