% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pheno.R
\name{import_amrrules_predictions}
\alias{import_amrrules_predictions}
\title{Import phenotype predictions from AMRrules output}
\usage{
import_amrrules_predictions(
  input,
  sample_col = "sample",
  species_col = "organism",
  ab_col = "drug",
  sir_col = "clinical category",
  ecoff_col = "phenotype",
  method = "genotyping",
  platform = "AMRfinderplus + AMRrules"
)
}
\arguments{
\item{input}{A string representing a dataframe, or a path to an input file, containing the AMRrules output "genome_summary" file, which should be a long-form TSV file with one row per sample and drug.}

\item{sample_col}{(optional, default \code{"sample"}) String indicating the name of the input data column that provides the sample name.}

\item{species_col}{(optional, default \code{"organism"}) String indicating the name of the input data column that provides a species name. If provided, this column will be converted to micro-organism class \code{mo} via \code{\link[AMR:as.mo]{AMR::as.mo()}}. If the \code{rename_cols} parameter is set to \code{TRUE}, this column will also be renamed as \code{spp_pheno}. If interpretation is switched on, this column will be used to identify the appropriate  breakpoints for interpretation of each row in the data table.}

\item{ab_col}{(optional, default \code{"drug"}) String indicating the name of the input data column that provides a drug name.}

\item{sir_col}{(optional, default \code{"clinical category"}) String indicating the name of the input data column that indicates the S/I/R prediction.}

\item{ecoff_col}{(optional, default \code{"phenotype"}) String indicating the name of the input data column that indicates the WT/NWT prediction.}

\item{method}{(optional, default \code{"genotyping"}) String indicating the value to record in a new \code{method} field added to the output table.}

\item{platform}{(optional, default \code{"AMRfinderplus + AMRrules"}) String indicating the value to record in a new \code{platform} field added to the output table.}
}
\value{
A data frame with the processed AST data, including additional columns:
}
\description{
This imports antimicrobial phenotype predictions (S/I/R and WT/NWT) generated from genotype data using \href{https://www.amrrules.org}{AMRrules}.
}
\examples{
\dontrun{
# import and process AST data from EBI, write formatted data to file for later use
predictions <- import_amrrules_predictions("Ecoli_genome_summary.tsv")
}
}
