% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ebi.R
\name{import_ebi}
\alias{import_ebi}
\title{Download antimicrobial phenotype data from the EBI AMR Portal}
\usage{
import_ebi(user_genus = NULL, user_release = NULL, user_antibiotic_name = NULL)
}
\arguments{
\item{user_genus}{String specifying a bacterial genus to download data for (default NULL, will pull all taxa)}

\item{user_release}{String specifying the data release to download (default NULL, will pull latest release)}

\item{user_antibiotic_name}{String specifying an antibiotic to download data for (default NULL, will pull all antibiotics)}
}
\value{
A data frame containing EBI genotype data
}
\description{
Download antimicrobial phenotype data from the EBI AMR Portal
}
\examples{
\dontrun{
import_ebi(
    user_genus="Salmonella",
    user_release="2025-12",
    user_antibiotic_name="ampicillin"
)
}
}
\keyword{internal}
