% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pheno.R
\name{import_ebi_ast_ftp}
\alias{import_ebi_ast_ftp}
\title{Import and Process AST Data files retrieved from the EBI AMR portal FTP site}
\usage{
import_ebi_ast_ftp(
  input,
  interpret_eucast = FALSE,
  interpret_clsi = FALSE,
  interpret_ecoff = FALSE
)
}
\arguments{
\item{input}{A string representing the input dataframe, or a path to an input file, to be processed.}

\item{interpret_eucast}{A logical value (default is FALSE). If \code{TRUE}, the function will interpret the susceptibility phenotype (SIR) for each row based on the MIC or disk diffusion values, against EUCAST human breakpoints. These will be reported in a new column \code{pheno_eucast}, of class 'sir'.}

\item{interpret_clsi}{A logical value (default is FALSE). If \code{TRUE}, the function will interpret the susceptibility phenotype (SIR) for each row based on the MIC or disk diffusion values, against CLSI human breakpoints. These will be reported in a new column \code{pheno_clsi}, of class 'sir'.}

\item{interpret_ecoff}{A logical value (default is FALSE). If \code{TRUE}, the function will interpret the wildtype vs nonwildtype status for each row based on the MIC or disk diffusion values, against epidemiological cut-off (ECOFF) values. These will be reported in a new column \code{ecoff}, of class 'sir' and coded as 'R' (nonwildtype) or 'S' (wildtype).}

\item{rename_cols}{A logical value (default is TRUE). If \code{TRUE}, the function will rename the provided columns (specified by \code{ab_col}, \code{mic_col}, \code{disk_col}, \code{species_col}, \code{id_col}) to the default names expected by AMRgen functions ('drug_agent', 'mic', 'disk', 'spp_pheno', 'id'), to match those output by the other \code{import_ast()} functions.}
}
\value{
A data frame with the processed AST data, including additional columns:
}
\description{
This function will import antibiotic susceptibility testing (AST) data suitable for downstream use with AMRgen analysis functions. The expected input is phenotype data retrieved from the EBI AMR Portal FTP site (ftp://ftp.ebi.ac.uk/pub/databases/amr_portal/releases/) either directly or via the function \code{download_ebi()}.
Note that files downloaded from the EBI AMR Portal web browser (https://www.ebi.ac.uk/amr/data/?view=experiments) are formatted differently and can be imported with the function \code{import_ebi_ast()}
}
\examples{
\dontrun{
# download Salmonella phenotype data from EBI
pheno_salmonella <- download_ebi(genus="Salmonella")

# reformat to simplify use with AMRgen functions
pheno_salmonella <- import_ebi_ast_ftp(pheno_salmonella)
}
}
