% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_geno.R
\name{import_geno}
\alias{import_geno}
\title{Import and process antimicrobial genotype data from common sources}
\usage{
import_geno(input, format = "amrfp", ...)
}
\arguments{
\item{input}{A string representing a dataframe, or a path to an input file, containing the phenotype data in a supported format. These files may be downloaded from public sources such EBI or NCBI, or the files may be generated using common bioinformatics software for AMR genotyping.}

\item{format}{A string indicating the format of the data: \code{"amrfp"} (default), \code{"ebi_web"}, \code{"ebi_ftp"}, \code{"kleborate"}, \code{"abricate"}, \code{"rgi"}. This determines which importer function the data is passed on to for processing (see below).}

\item{...}{Format-specific arguments. See
\itemize{
\item \code{"amrfp"} : \code{\link[=import_amrfp]{import_amrfp()}} AMRFinderPlus output
\item \code{"ebi_web"} : \code{\link[=import_amrfp_ebi_ftp]{import_amrfp_ebi_ftp()}} EBI-processed AMRFinderPlus results (from \href{https://ftp.ebi.ac.uk/pub/databases/amr_portal/releases/}{FTP} or \code{\link[=download_ebi]{download_ebi()}})
\item \code{"ebi_ftp"} : \code{\link[=import_amrfp_ebi_web]{import_amrfp_ebi_web()}} EBI-processed AMRFinderPlus results, (from \href{https://ebi.ac.uk/amr}{EBI AMR portal})
\item \code{"kleborate"} : \code{\link[=import_kleborate]{import_kleborate()}} Kleborate output
\item \code{"abricate"} : \code{\link[=import_abricate]{import_abricate()}} ABRicate output
\item \code{"rgi"} : \code{\link[=import_rgi]{import_rgi()}} CARD RGI output
}}
}
\value{
A data frame with the processed genotype data, with harmonised gene names, mapped drug agents, and drug classes which can be used for other functions of the ARMgen package:
\itemize{
\item \code{id}: The sample identifier (\code{character}).
\item \code{marker}: The name of the genotype marker as it appears in the input (e.g. \code{gyrA_S83F}) (\code{character}).
\item \code{gene}: The gene identifier (\code{character}).
\item \code{mutation}: The mutation detected within the gene, converted to \href{https://hgvs-nomenclature.org/stable/}{HGVS nomenclature} syntax (e.g. \code{Ser83Phe}) (\code{character}).
\item \code{node}: (for AMRFinderPlus input only) The node in the NCBI reference gene hierarchy corresponding to the gene (\code{character}).
\item \code{drug_class}: Name of the antibiotic group associated with the genotype marker, compatible with AMR pkg (\code{character}).
\item \code{drug}: Name of the specific antibiotic associated with the genotype marker, compatible with AMR pkg (\code{ab}). Value \code{NA} is assigned when the markers are annotated with a class only and not a specific antibiotic.
\item \verb{variation type}: (for AMRFinderPlus, ABRicate, or Kleborate results) Type of variation, e.g. \verb{Gene presence detected}, \verb{Protein variant detected}, \verb{Nucleotide variant detected}, \verb{Inactivating mutation detected}, \verb{Promoter variant detected}.
... Other fields specific to the input file
}
}
\description{
This function imports AMR genotyping datasets in formats generated by common bioinformatics tools (AMRFinderPlus, ABRicate, Kleborate, CARD RGI) as well as processed AMRFinderPlus downloadable from EBI.
Drug/class annotations given for each genotype marker in the input file are parsed to standard antibiotic names and/or antibiotic groups recognised by the AMR pkg, to facilitate extracting relevant genotype markers for comparison to phenotype data for a specific antibiotic (e.g. using \code{AMRgen} functions \code{\link[=get_binary_matrix]{get_binary_matrix()}}, \code{\link[=ppv]{ppv()}}, \code{\link[=amr_upset]{amr_upset()}} and \code{\link[=amr_logistic]{amr_logistic()}}).
}
\examples{
# Import AMRFinderPlus data file
data(ecoli_geno_raw)
head(ecoli_geno_raw)
geno <- import_geno(ecoli_geno_raw \%>\% head(n = 10), format = "amrfp")
head(geno)

\dontrun{
# Import ABRicate results (that were run using the default db, resfinder)
abricate_resfinder <- import_geno("path/to/abricate_resfinder.tsv",
  format = "abricate"
)

# Import ABRicate results that were run using an alternative db (ncbi)
abricate_ncbi <- import_geno("path/to/abricate_ncbi.tsv",
  format = "abricate",
  db = "ncbi"
)

# Import Kleborate results
kleborate_geno <- import_geno(kleborate_raw \%>\% head(n = 10),
  format = "kleborate"
)

# Import Kleborate results run with an older version without HGVS syntax
kleborate_old <- import_geno(kleborate_raw_v313 \%>\% head(n = 10),
  format = "kleborate",
  hgvs = FALSE
)

# Import CARD RGI results with default parameters
rgi_geno <- import_geno(rgi_EuSCAPE_raw \%>\% head(n = 10),
  format = "rgi"
)

# Import CARD RGI results with additional options
rgi_geno <- import_geno(rgi_raw \%>\% head(n = 10),
  format = "rgi",
  sample_id_sep = "_genomic.fna.txt:", exclude_loose = FALSE
)

# Download quinolone-related genotype data for E. coli, from EBI
ebi_geno_raw <- download_ebi(
  data = "genotype", species = "Escherichia coli",
  geno_subclass = "QUINOLONE"
)
# import the downloaded data
ebi_geno <- import_geno(ebi_geno_raw,
  format = "ebi_ftp"
)

# Download data from EBI web portal manually, and import the file
ebi_geno_from_web <- import_geno("amr_records.csv",
  format = "ebi_web"
)

# Download carbapenem-related genotype data for K. pneumoniae, from NCBI
ncbi_geno_raw <- query_ncbi_bq_geno(
  taxgroup = "Klebsiella pneumoniae",
  geno_subclass = "CARBAPENEM"
)
# import the downloaded data
geno <- import_geno(ncbi_geno_raw,
  format = "amrfp",
  sample_col = "biosample_acc"
)
}
}
