% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtdb.R
\name{import_gtdb}
\alias{import_gtdb}
\title{Import GTDB Output}
\usage{
import_gtdb(file = NULL, tbl = NULL, species_column = "Species")
}
\arguments{
\item{file}{File path to GTDB output file (tab-separated). If not given, tbl must be given.}

\item{tbl}{Data frame containing GTDB output. If not given, file must be given.}

\item{species_column}{Name of column containing the species call (default "Species").}
}
\value{
tbl Data frame containing GTDB output with AMR-parsed microorganism code (gtdb.mo) and species name (gtdb.species) appended.
}
\description{
Import GTDB output (from file or data frame) and parse the species name into a microorganism recognised by AMR package. AMR mo code and species name will be added to the data frame.
}
\examples{
import_gtdb(tbl = data.frame(Species = c("Pseudomonas_E piscis","Haemophilus_D parainfluenzae_A", "Acinetobacter calcoaceticus_C")))

}
