% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_kleborate.R
\name{import_kleborate}
\alias{import_kleborate}
\title{Import and Process Kleborate Results}
\usage{
import_kleborate(
  input_table,
  sample_col = "strain",
  kleborate_class_table = kleborate_classes,
  hgvs = TRUE
)
}
\arguments{
\item{input_table}{A character string specifying a dataframe or path to the Kleborate results table (TSV format).}

\item{sample_col}{A character string specifying the column that identifies samples in the dataset (default \code{strain}).}

\item{kleborate_class_table}{A tibble containing a reference table mapping Kleborate drug class column names (\code{Kleborate_Class}) to standardised drug classes (\code{drug_class}). Defaults to \code{kleborate_classes}, which is provided internally.}

\item{hgvs}{Logical indicating whether to expect mutations in HGVS format (used in Kleborate releases since v3.1.3). Default \code{TRUE}, which expects mutations formatted as e.g. "GyrA:p.S83F". Set to \code{FALSE} if your results were generated using older versions where mutations were formatted as e.g. "GyrA_83F".}
}
\value{
A tibble containing the processed AMR determinants and drug classes that is AMRgen compatible.
}
\description{
This function imports and processes genotyping results from Kleborate (https://github.com/klebgenomics/Kleborate), extracting antimicrobial resistance determinants and mapping them to standardised drug classes.
}
\details{
The function performs the following steps:
\itemize{
\item Reads the Kleborate output table.
\item Transforms Kleborate output into long form (i.e., one AMR determinant per row).
\item Maps Kleborate drug classes to standardised drug class names.
This processing ensures compatibility with downstream AMRgen analysis workflows.
}
}
\examples{
# example Kleborate data from EUSCAPE project
kleborate_raw

# import first few rows of this data frame and parse it as AMRfp data
kleborate_geno <- import_kleborate(kleborate_raw \%>\% head(n = 10), "strain")
}
