% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pheno.R
\name{import_pheno}
\alias{import_pheno}
\title{Import and process antimicrobial phenotype data from common sources}
\usage{
import_pheno(
  input,
  format = "ebi",
  interpret_eucast = FALSE,
  interpret_clsi = FALSE,
  interpret_ecoff = FALSE,
  ...
)
}
\arguments{
\item{input}{A string representing a dataframe, or a path to an input file, containing the phenotype data in a supported format. These files may be downloaded from public sources such as the \href{https://www.ebi.ac.uk/amr/data/?view=experiments}{EBI AMR web browser}, \href{ftp://ftp.ebi.ac.uk/pub/databases/amr_portal/releases/}{EBI FTP site}, or \href{https://www.ncbi.nlm.nih.gov/pathogens/ast}{NCBI browser}, or using the functions \code{\link[=download_ebi]{download_ebi()}}, \code{\link[=download_ncbi_ast]{download_ncbi_ast()}}, or \code{\link[=query_ncbi_bq_geno]{query_ncbi_bq_geno()}}; or the files may be exported from supported AST instruments.}

\item{format}{A string indicating the format of the data: \code{"ebi"} (default), \code{"ebi_web"}, \code{"ebi_ftp"}, \code{"ncbi"}, \code{"ncbi_biosample"}, \code{"vitek"}, \code{"microscan"}, \code{"phoenix"}, \code{"sensititre"}, or \code{"whonet"}. This determines which importer function the data is passed on to for processing (see below).}

\item{interpret_eucast}{A logical value (default is \code{FALSE}). If \code{TRUE}, the function will interpret the susceptibility phenotype (SIR) for each row based on the MIC or disk diffusion values, against EUCAST human breakpoints. These will be reported in a new column \code{pheno_eucast}, of class 'sir'.}

\item{interpret_clsi}{A logical value (default is \code{FALSE}). If \code{TRUE}, the function will interpret the susceptibility phenotype (SIR) for each row based on the MIC or disk diffusion values, against CLSI human breakpoints. These will be reported in a new column \code{pheno_clsi}, of class 'sir'.}

\item{interpret_ecoff}{A logical value (default is \code{FALSE}). If \code{TRUE}, the function will interpret the wildtype vs nonwildtype status for each row based on the MIC or disk diffusion values, against epidemiological cut-off (ECOFF) values. These will be reported in a new column \code{ecoff}, of class 'sir' and coded as \code{NWT} (nonwildtype) or \code{WT} (wildtype).}

\item{...}{Format-specific arguments. See
\itemize{
\item \code{"ebi"} : \code{\link[=import_ebi_ast]{import_ebi_ast()}}
\item \code{"ebi_web"} : \code{\link[=import_ebi_ast]{import_ebi_ast()}}
\item \code{"ebi_ftp"} :\code{\link[=import_ebi_ast_ftp]{import_ebi_ast_ftp()}}
\item \code{"ncbi"} : \code{\link[=import_ncbi_ast]{import_ncbi_ast()}}
\item \code{"ncbi_biosample"} : \code{\link[=import_ncbi_biosample]{import_ncbi_biosample()}}
\item \code{"vitek"} : \code{\link[=import_vitek_ast]{import_vitek_ast()}}
\item \code{"microscan"} : \code{\link[=import_microscan_ast]{import_microscan_ast()}}
\item \code{"sensititre"} : \code{\link[=import_sensititre_ast]{import_sensititre_ast()}}
\item \code{"phoenix"} : \code{\link[=import_phoenix_ast]{import_phoenix_ast()}}
\item \code{"whonet"} : \code{\link[=import_whonet_ast]{import_whonet_ast()}}
}}
}
\value{
A data frame with the processed AST data, including additional columns:
\itemize{
\item \code{id}: The sample identifier (\code{character}).
\item \code{spp_pheno}: The species phenotype, formatted using the \code{\link[AMR:as.mo]{AMR::as.mo()}} function (class \code{mo}).
\item \code{drug_agent}: The antibiotic used in the test, formatted using the \code{\link[AMR:as.ab]{AMR::as.ab()}} function (class \code{ab}).
\item \code{mic}: The minimum inhibitory concentration (MIC) value, formatted using the \code{\link[AMR:as.mic]{AMR::as.mic()}} function (class \code{mic}).
\item \code{disk}: The disk diffusion measurement (in mm), formatted using the \code{\link[AMR:as.disk]{AMR::as.disk()}} function (class \code{disk}).
\item \code{method}: The AST method (e.g., \code{"broth dilution"}, \code{"disk diffusion"}, \code{"Etest"}, \code{"agar dilution"}). Expected values are based on the NCBI/EBI antibiogram specification (\code{character}).
\item \code{platform}: The AST platform/instrument (e.g., \code{"Vitek"}, \code{"Phoenix"}, \code{"Sensititre"}) (\code{character}).
\item \code{guideline}: The AST standard recorded in the input file as being used for the AST assay (\code{character}).
\item \code{pheno_eucast}: The phenotype newly interpreted against EUCAST human breakpoint standards (as \code{S/I/R}), based on the MIC or disk diffusion data (class \code{sir}).
\item \code{pheno_clsi}: The phenotype newly interpreted against CLSI human breakpoint standards (as \code{S/I/R}), based on the MIC or disk diffusion data (class \code{sir}).
\item \code{ecoff}: The phenotype newly interpreted against the ECOFF (as \code{WT/NWT}), based on the MIC or disk diffusion data (class \code{sir}).
\item \code{pheno_provided}: The original phenotype interpretation provided in the input file, formatted using \code{\link[AMR:as.sir]{AMR::as.sir()}} (class \code{sir}).
\item \code{source}: The source of each data point (from the publications or bioproject field in the input file, or replaced with a single value passed in as the \code{source} parameter) (\code{character}).
}
}
\description{
This function imports an antibiotic susceptibility testing datasets in formats exported by EBI, NCBI, WHOnet and several automated AST instruments (Vitek, Microscan, Sensititre, Phoenix). It optionally can use the AMR package to interpret susceptibility phenotype (SIR) based on EUCAST or CLSI guidelines (human breakpoints and/or ECOFF).
}
\examples{
\dontrun{
# import NCBI data retrieved from Google Cloud, without re-interpreting resistance
head(staph_ast_ncbi_cloud_raw)
pheno <- import_pheno(staph_ast_ncbi_cloud_raw, format = "ncbi")

# import NCBI data where biosample column has been renamed to 'id'
head(staph_ast_ncbi_raw)
import_pheno(staph_ast_ncbi_raw, "ncbi", sample_col = "id")

# import NCBI data and re-interpret resistance (S/I/R) and WT/NWT (vs ECOFF)
head(ecoli_ast_raw)
pheno <- import_pheno(ecoli_ast_raw,
  format = "ncbi",
  interpret_eucast = TRUE, interpret_ecoff = TRUE
)

# download Klebsiella quasipneumoniae phenotype data from NCBI BioSample
kquasi_raw_ncbi <- download_ncbi_ast("Klebsiella quasipneumoniae")
head(kquasi_raw_ncbi)
# import the data and interpret against EUCAST breakpoints
pheno <- import_pheno(kquasi_raw_ncbi,
  format = "ncbi_biosample",
  interpret_eucast = T
)

# download Klebsiella quasipneumoniae phenotype data from EBI
kquasi_raw_ebi <- download_ebi(species = "Klebsiella quasipneumoniae")
head(kquasi_raw_ebi)
# import the data and interpret against ecoff
pheno <- import_pheno(kquasi_raw_ebi,
  format = "ebi_ftp",
  interpret_ecoff = TRUE
)

# import Vitek data from file, with default parameters
pheno <- import_pheno("vitek_export.tsv",
  format = "vitek"
)

# import Vitek data from file
# specify guideline that was used, remove dates, ignore expertized calls
pheno <- import_pheno("vitek_export.tsv",
  format = "vitek",
  instrument_guideline = "EUCAST 2025",
  use_expertized = FALSE,
  include_dates = FALSE
)
}
}
