% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rgi.R
\name{import_rgi}
\alias{import_rgi}
\title{Import and Process Resistance Gene Identifier (RGI) Results}
\usage{
import_rgi(
  input_table,
  orf_id_col = "ORF_ID",
  sample_id_sep = ".fasta.txt:",
  model_col = "Model_type",
  antibiotic_col = "Antibiotic",
  class_col = "Drug Class",
  exclude_loose = TRUE,
  rgi_short_name = rgi_short_name_table,
  rgi_drugs = rgi_drugs_table
)
}
\arguments{
\item{input_table}{A character string specifying a dataframe or path to the RGI results table (TSV format).}

\item{orf_id_col}{A character string specifying the column that identifies open reading frame ID (ORF_ID) in the dataset (default \code{ORF_ID}). This column includes the sample ID and the contig / genomic location and is a default output of RGI.}

\item{sample_id_sep}{A character string specifying the separator by which the sample ID is separated from the remaining text in \code{ORF_ID} (Default: \verb{.fasta.txt:}) . For example: in the \code{ORF_ID} column, "SAMEA3498968.fasta.txt:1_96 # 109511 # 110635....", the sample ID separator is \verb{.fasta.txt:}.}

\item{model_col}{A character string specifying the column that identifies model type identified by RGI (default \code{Model_type}).}

\item{antibiotic_col}{Character string specifying the antibiotic column (default \code{Antibiotic}).}

\item{class_col}{Character string specifying the drug class column (default \verb{Drug Class}).}

\item{exclude_loose}{Logical indicating whether to exclude Loose hits (AMR markers that fall below a curated bitscore cutoff as defined by CARD/RGI). Default \code{TRUE}, which excludes Loose hits.}

\item{rgi_short_name}{A tibble containing a reference table mapping model IDs (from CARD/RGI) to shortened model names as provided by CARD (https://card.mcmaster.ca/download in aro_index.tsv). Defaults to \code{rgi_short_name_table}, which is provided internally.}

\item{rgi_drugs}{A tibble containing a reference table mapping CARD drug class / drug agents to standardised drug classes/names. Defaults to \code{rgi_drugs_table}, which is provided internally.}
}
\value{
A tibble containing the processed AMR determinants and drug classes that is AMRgen compatible. The output retains the original columns from the RGI output along with the newly mapped variables.
}
\description{
This function imports and processes genotyping results from the Resistance Gene Identifier (RGI, https://github.com/arpcard/rgi), extracting antimicrobial resistance determinants and mapping them to standardised drug classes/antibiotics.
}
\details{
The function performs the following steps:
\itemize{
\item Reads the RGI output table.
\item Transforms RGI output into long form (i.e., one AMR determinant AND drug class / antibiotic per row).
\item Maps CARD drug classes and antibiotics to standardised names.
This processing ensures compatibility with downstream AMRgen analysis workflows.
}
}
\examples{
# example RGI data (including Perfect, Strict, and Loose hits)
rgi_raw

# import using sample_id_sep=`_genomic.fna.txt:` and include Loose hits
rgi <- import_rgi(rgi_raw, sample_id_sep = "_genomic.fna.txt:", exclude_loose = FALSE)

# example RGI data from EuSCAPE project (including only Perfect and Strict hits)
rgi_EuSCAPE_raw

# import using defaults (sample_id_sep=`.fasta.txt:`, exclude_loose = `TRUE`)
import_rgi(rgi_EuSCAPE_raw)
}
