% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pheno.R
\name{import_vitek_pheno}
\alias{import_vitek_pheno}
\title{Import and process antimicrobial phenotype data exported from Vitek instruments}
\usage{
import_vitek_pheno(
  input,
  sample_col = "Lab ID",
  source = NULL,
  species = NULL,
  ab = NULL,
  instrument_guideline = NULL,
  use_expertized = TRUE,
  interpret_eucast = FALSE,
  interpret_clsi = FALSE,
  interpret_ecoff = FALSE,
  include_dates = TRUE
)
}
\arguments{
\item{input}{A dataframe or path to a CSV file containing Vitek AST output data}

\item{sample_col}{String indicating the name of the column containing sample identifiers. Default: \code{"Lab ID"}}

\item{source}{Optional string value to record in the \code{source} column for all data points (e.g., dataset name or study identifier)}

\item{species}{Optional string indicating a single species to use for phenotype interpretation (otherwise this is inferred per-sample from the input)}

\item{ab}{Optional string indicating a single antibiotic to use for phenotype interpretation (otherwise this is inferred per-sample from the input)}

\item{instrument_guideline}{Optional string indicating the guideline used by the instrument for SIR interpretation (e.g., "EUCAST 2025", "CLSI 2025"), used to record the \code{guideline} in the output file. Default: \code{NULL}}

\item{use_expertized}{Use expertized SIR (\code{TRUE}, default) or instrument SIR (\code{FALSE})}

\item{interpret_eucast}{A logical value (default is \code{FALSE}). If \code{TRUE}, the function will re-interpret the susceptibility phenotype (SIR) for each observation based on the MIC values, against EUCAST human breakpoints. These will be reported in a new column \code{pheno_eucast}, of class \code{sir}.}

\item{interpret_clsi}{A logical value (default is \code{FALSE}). If \code{TRUE}, the function will re-interpret the susceptibility phenotype (SIR) for each observation based on the MIC values, against CLSI human breakpoints. These will be reported in a new column \code{pheno_clsi}, of class \code{sir}.}

\item{interpret_ecoff}{A logical value (default is \code{FALSE}). If \code{TRUE}, the function will re-interpret the wildtype vs nonwildtype status for each observation based on the MIC values, against epidemiological cut-off (ECOFF) values. These will be reported in a new column \code{ecoff}, of class \code{sir} and coded as \code{NWT} (nonwildtype) or \code{WT} (wildtype).}

\item{include_dates}{Logical indicating whether to include \code{collection_date} and \code{testing_date} in output. Default: \code{TRUE}}
}
\value{
Standardised AST data frame
}
\description{
This function imports antimicrobial susceptibility testing (AST) data from Vitek instrument output files (wide CSV format)
and converts it to the standardised long-format used by AMRgen.
}
