% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pheno.R
\name{interpret_ast}
\alias{interpret_ast}
\title{Interpret AST data in a standard format tibble}
\usage{
interpret_ast(
  ast,
  interpret_ecoff = TRUE,
  interpret_eucast = TRUE,
  interpret_clsi = TRUE,
  species = NULL,
  ab = NULL
)
}
\arguments{
\item{ast}{A tibble containing the AST measures in standard AMRgen format, as output by \code{import_ast}. It must contain assay measurements in columns 'mic' (class mic) and/or 'disk'. Interpretation requires an organism (column 'spp_pheno' of class 'mo', or a single value passed via the 'species' parameter) and an antibiotic (column 'drug_agent' of class 'ab', or a single value passed via the 'ab' parameter).}

\item{interpret_ecoff}{A logical value (default is FALSE). If \code{TRUE}, the function will interpret the wildtype vs nonwildtype status for each row based on the MIC or disk diffusion values, against epidemiological cut-off (ECOFF) values. These will be reported in a new column \code{ecoff}, of class 'sir' and coded as 'R' (nonwildtype) or 'S' (wildtype).}

\item{interpret_eucast}{A logical value (default is FALSE). If \code{TRUE}, the function will interpret the susceptibility phenotype (SIR) for each row based on the MIC or disk diffusion values, against EUCAST human breakpoints. These will be reported in a new column \code{pheno_eucast}, of class 'sir'.}

\item{interpret_clsi}{A logical value (default is FALSE). If \code{TRUE}, the function will interpret the susceptibility phenotype (SIR) for each row based on the MIC or disk diffusion values, against CLSI human breakpoints. These will be reported in a new column \code{pheno_clsi}, of class 'sir'.}

\item{species}{(optional) Name of the species to use for phenotype interpretation. By default, the spp_pheno field in the input file will be assumed to specify the species for each sample, but if this is missing or you want to override it in the interpretation step, you may provide a single species name via this parameter.}

\item{ab}{(optional) Name of the antibiotic to use for phenotype interpretation. By default, the drug_agent field in the input file will be assumed to specify the antibiotic for each sample, but if this is missing or you want to override it in the interpretation step, you may provide a single antibiotic name via this parameter.}
}
\value{
A copy of the input table, with additional columns:
\itemize{
\item \code{pheno_eucast}: The phenotype newly interpreted against EUCAST human breakpoint standards (as S/I/R), based on the MIC or disk diffusion data.
\item \code{pheno_clsi}: The phenotype newly interpreted against CLSI human breakpoint standards (as S/I/R), based on the MIC or disk diffusion data.
\item \code{ecoff}: The phenotype newly interpreted against the ECOFF (as S/R), based on the MIC or disk diffusion data.
\item \code{spp_pheno}: The species phenotype, formatted using the \code{as.mo} function (either taken from the input table, or the single value specified by 'species' parameter).
\item \code{drug_agent}: The antibiotic used in the test, formatted using the \code{as.ab} function (either taken from the input table, or the single value specified by 'ab' parameter)..
}
}
\description{
This function applies human EUCAST or CLSI breakpoints, and/or ECOFF, to interpret AST data.
}
\examples{
# import without re-interpreting resistance
pheno <- import_ncbi_ast(ecoli_ast_raw)
head(pheno)

# interpret phenotypes
pheno <- interpret_ast(pheno)

\dontrun{
pheno <- read_csv("AST.csv") \%>\%
  # convert antibiotic field to 'drug_agent' of class 'ab'
  mutate(drug_agent = as.ab(antibiotic)) \%>\%
  mutate(mic = paste0(sign, MIC)) \%>\%
  mutate(mic = as.mic(mic)) # create a single 'mic' column of class 'mic'

pheno <- interpret_ast(pheno, species = "Escherichia coli")
}
}
