% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pheno.R
\name{interpret_pheno}
\alias{interpret_pheno}
\title{Interpret antimicrobial susceptibility phenotype data in a standard format tibble}
\usage{
interpret_pheno(
  ast,
  interpret_ecoff = TRUE,
  interpret_eucast = TRUE,
  interpret_clsi = TRUE,
  species = NULL,
  ab = NULL
)
}
\arguments{
\item{ast}{A tibble containing the phenotype measures in standard AMRgen format, as output by \code{\link[=import_pheno]{import_pheno()}}. It must contain assay measurements in columns \code{mic} (class \code{mic}) and/or \code{disk} (class \code{disk}). Interpretation requires an organism (column \code{spp_pheno} of class \code{mo}, or a single value passed via the \code{species} parameter) and an antibiotic (column \code{drug} of class \code{ab}, or a single value passed via the \code{ab} parameter).}

\item{interpret_ecoff}{A logical value (default is \code{FALSE}). If \code{TRUE}, the function will interpret the wildtype vs nonwildtype status for each row based on the MIC or disk diffusion values, against epidemiological cut-off (ECOFF) values. These will be reported in a new column \code{ecoff}, of class \code{sir} and coded as \code{NWT} (nonwildtype) or \code{WT} (wildtype).}

\item{interpret_eucast}{A logical value (default is \code{FALSE}). If \code{TRUE}, the function will interpret the susceptibility phenotype (SIR) for each row based on the MIC or disk diffusion values, against EUCAST human breakpoints. These will be reported in a new column \code{pheno_eucast}, of class \code{sir}.}

\item{interpret_clsi}{A logical value (default is \code{FALSE}). If \code{TRUE}, the function will interpret the susceptibility phenotype (SIR) for each row based on the MIC or disk diffusion values, against CLSI human breakpoints. These will be reported in a new column \code{pheno_clsi}, of class \code{sir}.}

\item{species}{(optional) Name of the species to use for phenotype interpretation. By default, the \code{spp_pheno} field in the input file will be assumed to specify the species for each sample, but if this is missing or you want to override it in the interpretation step, you may provide a single species name via this parameter.}

\item{ab}{(optional) Name of the antibiotic to use for phenotype interpretation. By default, the \code{drug} field in the input file will be assumed to specify the antibiotic for each sample, but if this is missing or you want to override it in the interpretation step, you may provide a single antibiotic name via this parameter.}
}
\value{
A copy of the input table, with additional columns:
\itemize{
\item \code{pheno_eucast}: The phenotype newly interpreted against EUCAST human breakpoint standards (as S/I/R), based on the MIC or disk diffusion data.
\item \code{pheno_clsi}: The phenotype newly interpreted against CLSI human breakpoint standards (as S/I/R), based on the MIC or disk diffusion data.
\item \code{ecoff}: The phenotype newly interpreted against the ECOFF (as WT/NWT), based on the MIC or disk diffusion data.
\item \code{spp_pheno}: The species phenotype, formatted using the \code{\link[AMR:as.mo]{AMR::as.mo()}} function (either taken from field \code{spp_pheno} in the input table, or the single value specified by \code{species} parameter).
\item \code{drug}: The antibiotic used in the test, formatted using the \code{\link[AMR:as.ab]{AMR::as.ab()}} function (either taken from field \code{drug} in the input table, or the single value specified by \code{ab} parameter).
}
}
\description{
This function applies human EUCAST or CLSI breakpoints, and/or ECOFF, to interpret antimicrobial susceptibility testing (AST) data.
}
\examples{
# import without re-interpreting resistance
pheno <- import_ncbi_pheno(ecoli_pheno_raw)
head(pheno)

# interpret phenotypes
pheno <- interpret_pheno(pheno)

\dontrun{
pheno <- read_csv("AST.csv") \%>\%
  # convert antibiotic field to 'drug' of class 'ab'
  mutate(drug = as.ab(antibiotic)) \%>\%
  mutate(mic = paste0(sign, MIC)) \%>\%
  mutate(mic = as.mic(mic)) # create a single 'mic' column of class 'mic'

pheno <- interpret_pheno(pheno, species = "Escherichia coli")
}
}
