% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kleborate_classes}
\alias{kleborate_classes}
\title{Table mapping Kleborate drug class columns}
\format{
\code{kleborate_classes} A data frame with 21 rows and 2 columns:
\itemize{
\item \code{Kleborate_Class}: Column name in Kleborate output files
\item \code{drug_class}: Valid drug class name recognised by AMR pkg
}
}
\usage{
kleborate_classes
}
\description{
Table mapping Kleborate drug class columns to class names recognised by AMR pkg
}
\keyword{datasets}
