% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kleborate_raw_v313}
\alias{kleborate_raw_v313}
\title{Example Kleborate v3.1.3 Genotype Data from EuSCAPE project}
\format{
\code{kleborate_raw_v313} A data frame with 1,490 rows and 113 columns:
\itemize{
\item \code{strain}: Sample identifier
\item ...: Kleborate results columns
}
}
\source{
ENA BioProject \href{https://www.ebi.ac.uk/ena/browser/view/PRJEB10018}{PRJEB10018}.
See David \emph{et al.} (2019) \url{https://doi.org/10.1038/s41564-019-0492-8}.
}
\usage{
kleborate_raw_v313
}
\description{
Raw Kleborate v3.1.3 results file for \emph{Klebsiella pneumoniae} genomes, one row per sample.
}
\keyword{datasets}
