% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kp_mero_amrfp}
\alias{kp_mero_amrfp}
\title{Example AMRFinderPlus Genotype Data from EuSCAPE project}
\format{
\code{kp_mero_amrfp} A data frame with 32,385 rows and 34 columns:
\itemize{
\item \code{id}:  BioSample.
\item \code{drug_agent}, \code{drug_class}: Antibiotic agent and class, determined by parsing AMRFinderPlus \code{subclass} field in the downloaded file.
\item \code{gene}, \code{node}, \code{marker}: gene identifiers.
\item \code{mutation}: mutation within gene, parsed into HGVS nomenclature format from \code{amr_element_symbol} field in the downloaded file.
\item \verb{\% Coverage of reference}:  \% Coverage of reference.
\item \verb{\% Identity to reference}:  \% Identity to reference.
\item ...: Additional data columns from AMRFinderPlus
}
}
\source{
\href{https://www.ebi.ac.uk/amr}{EBI AMR Portal}.
See David \emph{et al.} (2019) \url{https://doi.org/10.1038/s41564-019-0492-8}.
}
\usage{
kp_mero_amrfp
}
\description{
AMRFinderPlus results file for Klebsiella pneumoniae from EuSCAPE project, one AMR determinant per row, downloaded from the EBI AMR portal using \code{\link[=download_ebi]{download_ebi()}} and imported using \code{\link[=import_geno]{import_geno()}}.
}
\keyword{datasets}
