% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kp_mero_euscape}
\alias{kp_mero_euscape}
\title{Meropenem Phenotype Data from EuSCAPE project}
\format{
\code{kp_mero_euscape} A data frame with 1,490 rows and 43 columns:
\itemize{
\item \code{id}: Sample identifier, imported from the \code{BioSample} column in the raw input.
\item \code{drug}: Antibiotic code, interpreted from \code{Antibiotic} using \code{as.ab}.
\item \code{mic}: Minimum inhibitory concentration, formatted using \code{as.mic}.
\item \code{disk}: Disk diffusion zone, formatted using \code{as.disk}.
\item \code{method}, \code{platform}, \code{guideline}: Test method and platform and interpretation guideline.
\item \code{pheno_provided}: S/I/R interpretation as provided in the raw input.
\item \code{spp_pheno}: Species identifier, interpreted from \verb{Scientific name} using \code{as.mo}, used to interpret \code{ecoff} and \code{pheno} columns.
\item ...: Additional data columns from EBI AMR Portal
}
}
\source{
\href{https://www.ebi.ac.uk/amr}{EBI AMR Portal}.
See David \emph{et al.} (2019) \url{https://doi.org/10.1038/s41564-019-0492-8}.
}
\usage{
kp_mero_euscape
}
\description{
Meropenem phenotype data for Klebsiella pneumoniae from EuSCAPE project, one sample per row, downloaded from the EBI AMR portal using \code{\link[=download_ebi]{download_ebi()}} and imported using \code{\link[=import_pheno]{import_pheno()}}.
}
\keyword{datasets}
