% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{logistf_details}
\alias{logistf_details}
\title{Extract Details from a logistf Model}
\usage{
logistf_details(model)
}
\arguments{
\item{model}{A fitted logistf model object.}
}
\value{
A tibble containing the estimates, confidence intervals, and p-values for each predictor in the model.

Example
library(logistf)
model <- logistf(R ~ ., data=dat)
model_details <- logistf_details(model)
autoplot(model_details)
}
\description{
This function extracts and formats the estimates, confidence intervals, and p-values from a fitted logistf model.
}
