% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{merge_logreg_soloppv}
\alias{merge_logreg_soloppv}
\title{Merge Logistic Regression and Solo PPV Statistics}
\usage{
merge_logreg_soloppv(model, solo_stats, title = NULL)
}
\arguments{
\item{model}{A data frame containing logistic regression model statistics.}

\item{solo_stats}{A data frame containing solo PPV statistics.}

\item{title}{An optional title for the plot.}
}
\value{
A list containing:
\itemize{
\item \code{combined}: A merged data frame of logistic regression and PPV statistics.
\item \code{plot}: A ggplot object showing the relationship between model estimates and PPV.
}
}
\description{
This function merges logistic regression model statistics with solo PPV statistics and creates a combined plot.
}
\examples{
\dontrun{
soloPPV_cipro <- solo_ppv_analysis(ecoli_geno, ecoli_ast,
  antibiotic = "Ciprofloxacin",
  drug_class_list = c("Quinolones"),
  sir_col = "pheno"
)
logistic_cipro <- amr_logistic(ecoli_geno, ecoli_ast,
  "Ciprofloxacin", c("Quinolones"),
  maf = 5
)
allstatsR <- merge_logreg_soloppv(logistic_cipro$modelR,
  soloPPV_cipro$solo_stats \%>\% filter(category == "R"),
  title = "Quinolone markers vs Cip R"
)
}
}
