% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MICROBIGGE_CATB3}
\alias{MICROBIGGE_CATB3}
\title{MicroBIGG-E for E.coli containing catB3}
\format{
\code{MICROBIGGE_CATB3} A data frame with 35054 rows and 18 columns:
\itemize{
\item \verb{#Scientific name}:  #Scientific name.
\item \code{Protein}:  Protein.
\item \code{BioSample}:  BioSample.
\item \code{Isolate}:  Isolate.
\item \code{Contig}:  Contig.
\item \code{Start}:  Start.
\item \code{Stop}:  Stop.
\item \code{Strand}:  Strand.
\item \verb{Element symbol}:  Element symbol.
\item \verb{Element name}:  Element name.
\item \code{Type}:  Type.
\item \code{Scope}:  Scope.
\item \code{Subtype}:  Subtype.
\item \code{Class}:  Class.
\item \code{Subclass}:  Subclass.
\item \code{Method}:  Method.
\item \verb{\% Coverage of reference}:  \% Coverage of reference.
\item \verb{\% Identity to reference}:  \% Identity to reference.
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pathogens/microbigge/#catB3\%20AND\%20Escherichia}
}
\usage{
MICROBIGGE_CATB3
}
\description{
NCBI Pathogen Detection Microbial Browser for Identification of Genetic and Genomic Elements (MicroBIGG-E)
for Escherichia coli containing catB3
}
\keyword{datasets}
