% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MICROBIGGE_Ecoli_CHLR}
\alias{MICROBIGGE_Ecoli_CHLR}
\title{MicroBIGG-E for E.coli containing all chloramphenicol resistance genes}
\format{
\code{MICROBIGGE_Ecoli_CHLR} A data frame with 95,776 rows and 27 columns:
\itemize{
\item \code{id}:  BioSample.
\item \code{drug}, \code{drug_class}: Antibiotic agent and class, determined by parsing AMRFinderPlus \code{subclass} field in the downloaded file.
\item \code{gene}, \code{node}, \code{marker}: gene identifiers.
\item \code{mutation}: mutation within gene, parsed into HGVS nomenclature format from \code{amr_element_symbol} field in the downloaded file.
\item \verb{\% Coverage of reference}:  \% Coverage of reference.
\item \verb{\% Identity to reference}:  \% Identity to reference.
\item ...: Additional data columns from AMRFinderPlus
#' @source \url{https://www.ncbi.nlm.nih.gov/pathogens/microbigge/#chloramphenicol\%20AND\%20Escherichia}
}
}
\usage{
MICROBIGGE_Ecoli_CHLR
}
\description{
Data from NCBI Pathogen Detection Microbial Browser for Identification of Genetic and Genomic Elements (MicroBIGG-E),
containing all chloramphenicol resistance gene hits in Escherichia coli genomes, imported using \code{\link[=import_amrfp]{import_amrfp()}}.
}
\keyword{datasets}
