% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCBI_AST_CHL}
\alias{NCBI_AST_CHL}
\title{NCBI AST for Escherichia coli tested against chloramphenicol}
\format{
\code{NCBI_AST_CHL} A data frame with 6859 rows and 17 columns:
\itemize{
\item \code{BioSample}:  BioSample.
\item \verb{Organism group}:  Organism group.
\item \verb{Scientific name}:  Scientific name.
\item \verb{Isolation type}:  Isolation type.
\item \code{Location}:  Location.
\item \verb{Isolation source}:  Isolation source.
\item \code{Isolate}:  Isolate.
\item \code{Antibiotic}:  Antibiotic.
\item \verb{Resistance phenotype}:  Resistance phenotype.
\item \verb{Measurement sign}:  Measurement sign.
\item \code{MIC (mg/L)}:  MIC (mg/L).
\item \verb{Disk diffusion (mm)}:  Disk diffusion (mm).
\item \verb{Laboratory typing platform}:  Laboratory typing platform.
\item \code{Vendor}:  Vendor.
\item \verb{Laboratory typing method version or reagent}:  Laboratory typing method version or reagent.
\item \verb{Testing standard}:  Testing standard.
\item \verb{Create date}:  Create date.
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pathogens/ast#chloramphenicol\%20AND\%20Escherichia}
}
\usage{
NCBI_AST_CHL
}
\description{
NCBI Antibiotic Susceptibility Test (AST) Browser data for Escherichia coli tested against chloramphenicol.
}
\keyword{datasets}
