% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ncbi_biosample.R, R/ncbi_biosample.R
\name{import_ncbi_biosample}
\alias{import_ncbi_biosample}
\alias{export_ncbi_biosample}
\title{Import/Export BioSample Antibiograms}
\usage{
import_ncbi_biosample(x)

export_ncbi_biosample(data, file, overwrite = FALSE)

import_ncbi_biosample(x)

export_ncbi_biosample(data, file, overwrite = FALSE)
}
\arguments{
\item{x}{data blah blah}

\item{data}{data set containing SIR results}

\item{file}{path to which biosample must be written to}

\item{overwrite}{overwrite current \code{file}?}
}
\description{
These functions allow to work with \href{https://www.ncbi.nlm.nih.gov/biosample/docs/antibiogram/}{NCBI BioSample Antibiograms}.

These functions allow to work with \href{https://www.ncbi.nlm.nih.gov/biosample/docs/antibiogram/}{NCBI BioSample Antibiograms}.
}
\examples{
# example code

# example code

}
