% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCBI_Ecoli_AST_chl}
\alias{NCBI_Ecoli_AST_chl}
\title{NCBI AST for Escherichia coli tested against chloramphenicol}
\format{
\code{NCBI_Ecoli_AST_chl} A data frame with 6,859 rows and 17 columns:
\itemize{
\item \code{id}: Sample identifier, imported from the \code{BioSample} column in the raw input.
\item \code{drug_agent}: Antibiotic code, interpreted from \code{Antibiotic} using \code{as.ab}.
\item \code{mic}: Minimum inhibitory concentration, formatted using \code{as.mic}.
\item \code{disk}: Disk diffusion zone, formatted using \code{as.disk}.
\item \code{method}, \code{platform}, \code{guideline}: Test method and platform and interpretation guideline.
\item \code{pheno_provided}: S/I/R interpretation as provided in the raw input.
\item \code{spp_pheno}: Species identifier, interpreted from \verb{Scientific name} using \code{as.mo}, used to interpret \code{ecoff} and \code{pheno} columns.
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pathogens/ast#chloramphenicol\%20AND\%20Escherichia}
}
\usage{
NCBI_Ecoli_AST_chl
}
\description{
NCBI Antibiotic Susceptibility Test (AST) Browser data for Escherichia coli tested against chloramphenicol.
}
\keyword{datasets}
