% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ngono_cro_geno_raw}
\alias{ngono_cro_geno_raw}
\title{N. gonorrhoeae PBP2 Mutations Genotype Data Use Case 2}
\format{
\code{ngono_cro_geno_raw} A data frame with 23,885 rows and 24 columns:
\itemize{
\item \code{Name}: Sample identifier (ENA run accession).
\item \verb{Protein id}: Protein identifier (all \code{NA}).
\item \verb{Contig id}: Assembly contig identifier.
\item \code{Start}, \code{Stop}: Start and stop positions of the AMR element on the contig.
\item \code{Strand}: Strand orientation (\code{+} or \code{-}).
\item \verb{Element symbol}: AMRFinderPlus element symbol (e.g. \code{pbp2}, \code{penA_A510V}).
\item \verb{Element name}: Full descriptive name of the AMR element.
\item \code{Scope}: AMRFinderPlus scope (\code{core} or \code{plus}).
\item \verb{Element Type}: Type of AMR element (e.g. \code{AMR}).
\item \code{Subtype}: Subtype of AMR element (e.g. \code{POINT}, \code{AMR}).
\item \code{Class}: Antibiotic class (e.g. \code{BETA-LACTAM}).
\item \code{Subclass}: Antibiotic subclass (e.g. \code{CEPHALOSPORIN}).
\item \code{Method}: Detection method (e.g. \code{POINTX}, \code{BLASTX}).
\item \verb{Target length}, \verb{Reference sequence length}, \verb{Alignment length}: Sequence length metrics (bp).
\item \verb{\% Coverage of reference}, \verb{\% Identity to reference}: Alignment quality metrics.
\item \verb{Closest reference accession}: NCBI accession of the closest reference sequence.
\item \verb{Closest reference name}: Name of the closest reference sequence.
\item \verb{HMM accession}, \verb{HMM description}: HMM-based annotation fields (all \code{NA}).
\item \verb{Hierarchy node}: Gene hierarchy node name required for \link{import_amrfp}.
}
}
\source{
ENA projects
\href{https://www.ebi.ac.uk/ena/browser/view/PRJEB58139}{PRJEB58139},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJEB57389}{PRJEB57389},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJEB76977}{PRJEB76977},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJEB45627}{PRJEB45627},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA577446}{PRJNA577446},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA776899}{PRJNA776899},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA778600}{PRJNA778600},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA560592}{PRJNA560592},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA874857}{PRJNA874857},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA909328}{PRJNA909328},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA957547}{PRJNA957547},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA1161034}{PRJNA1161034},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA1189294}{PRJNA1189294},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA1067895}{PRJNA1067895}.
See also Golparian \emph{et al.} (2024) \url{https://doi.org/10.1016/S2666-5247(23)00370-1},
Day \emph{et al.} (2022) \url{https://doi.org/10.2807/1560-7917.ES.2022.27.46.2200803},
Fifer \emph{et al.} (2024) \url{https://doi.org/10.1093/jac/dkae369},
van der Veen \emph{et al.} (2026) \url{https://doi.org/10.1093/cid/ciaf530},
Unemo \emph{et al.} (2024) \url{https://doi.org/10.1093/jac/dkae176}.
}
\usage{
ngono_cro_geno_raw
}
\description{
Raw concatenated output from AMRFinderPlus run on \emph{Neisseria gonorrhoeae} genome
assemblies from multiple studies enriched for ceftriaxone decreased susceptibility
and resistance (total n = 2,101 genomes). This object serves as the genotype input
for \link{import_amrfp} and is used to investigate mosaic PBP2 (\emph{penA}) variants
associated with ceftriaxone resistance.
}
\keyword{datasets}
