% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ngono_cro_pheno_raw}
\alias{ngono_cro_pheno_raw}
\title{N. gonorrhoeae PBP2 Mutations Phenotype Data Use Case 2}
\format{
\code{ngono_cro_pheno_raw} A data frame with 2,191 rows and 4 columns:
\itemize{
\item \code{id}: Sample identifier (ENA run accession).
\item \code{Ceftriaxone}: MIC value in mg/L for ceftriaxone (character, to preserve inequality prefixes).
\item \code{penA}: \emph{penA} mosaic allele type identifier (e.g. \code{60.001}, \code{237.001}).
\item \code{study}: Source study identifier (e.g. \code{fifer2024}).
}
}
\source{
ENA projects \href{https://www.ebi.ac.uk/ena/browser/view/PRJEB58139}{PRJEB58139},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJEB57389}{PRJEB57389},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJEB76977}{PRJEB76977},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJEB45627}{PRJEB45627},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA577446}{PRJNA577446},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA776899}{PRJNA776899},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA778600}{PRJNA778600},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA560592}{PRJNA560592},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA874857}{PRJNA874857},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA909328}{PRJNA909328},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA957547}{PRJNA957547},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA1161034}{PRJNA1161034},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA1189294}{PRJNA1189294},
\href{https://www.ebi.ac.uk/ena/browser/view/PRJNA1067895}{PRJNA1067895}.
See also Golparian \emph{et al.} (2024) \url{https://doi.org/10.1016/S2666-5247(23)00370-1},
Day \emph{et al.} (2022) \url{https://doi.org/10.2807/1560-7917.ES.2022.27.46.2200803},
Fifer \emph{et al.} (2024) \url{https://doi.org/10.1093/jac/dkae369},
van der Veen \emph{et al.} (2026) \url{https://doi.org/10.1093/cid/ciaf530},
Unemo \emph{et al.} (2024) \url{https://doi.org/10.1093/jac/dkae176}.
}
\usage{
ngono_cro_pheno_raw
}
\description{
Ceftriaxone MIC data for \emph{Neisseria gonorrhoeae} isolates from multiple studies
enriched for decreased susceptibility and resistance to ceftriaxone, including
Euro-GASP 2020, ceftriaxone-resistant isolates from the United Kingdom and England,
isolates from Asia, and WHO reference genomes. Used to investigate mosaic PBP2
variants and their association with ceftriaxone MICs.
}
\keyword{datasets}
