% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ngono_tet_pheno_raw}
\alias{ngono_tet_pheno_raw}
\title{N. gonorrhoeae Tetracycline Resistance Phenotype Data Use Case 3}
\format{
\code{ngono_tet_pheno_raw} A data frame with 409 rows and 2 columns:
\itemize{
\item \code{id}: Sample identifier (ENA run accession).
\item \code{Tetracycline}: MIC value in mg/L for tetracycline (numeric).
}
}
\source{
ENA project \href{https://www.ebi.ac.uk/ena/browser/view/PRJEB83795}{PRJEB83795}.
See Sánchez-Serrano \emph{et al.} (2026) \url{https://doi.org/10.1016/j.cmi.2025.12.026}.
}
\usage{
ngono_tet_pheno_raw
}
\description{
Tetracycline MIC data for 409 \emph{Neisseria gonorrhoeae} isolates collected in
Eastern Spain between 2021 and 2024, used to investigate genetic determinants
of tetracycline resistance and implications for doxy-PEP strategies.
Described in Sánchez-Serrano \emph{et al.} (2026) \url{doi:10.1016/j.cmi.2025.12.026}.
}
\keyword{datasets}
