% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno_eco_2075}
\alias{pheno_eco_2075}
\title{E. coli AST data from Mills et al 2022}
\format{
\code{pheno_eco_2075} A data frame with 37350 rows and 37 columns representing MIC results for 2075 E. coli isolates tested against 18 drug agents, using BD Phoenix.

Columns include:
\itemize{
\item \code{id}: Sample identifier (BioSample).
\item \code{drug_agent}: Antibiotic identifier, as class 'ab'.
\item \code{mic}: MIC data, as class 'mic'.
\item \code{pheno_provided}: S/I/R phenotypes as downloaded from EBI.
\item ...: Additional data columns from EBI.
}
}
\source{
\url{https://www.ebi.ac.uk/amr}
}
\usage{
pheno_eco_2075
}
\description{
Phenotyping data published in Mills et al, Genome Medicine (2022) 14:147, downloaded via EBI AMR portal, and imported to AMRgen phenotype table format. Corresponding genotype data is available in \code{geno_eco_2075}.
}
\keyword{datasets}
