% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{plot_combined_stats}
\alias{plot_combined_stats}
\title{Plot Combined Statistics}
\usage{
plot_combined_stats(combined_stats, sig = 0.05, title = NULL)
}
\arguments{
\item{combined_stats}{A data frame containing combined statistics from logistic regression and solo PPV.}

\item{sig}{A significance level for the logistic regression p-values. Default is 0.05.}

\item{title}{An optional title for the plot.}
}
\value{
A ggplot2 object representing the combined plot.
}
\description{
This function creates a plot of combined logistic regression and solo PPV statistics.
}
