% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resfinder}
\alias{resfinder}
\title{Resfinder Drug Classes and Agents Mapping Database}
\format{
A data frame (or tibble) containing mapping columns:
\describe{
\item{AMRFP_subclass}{AMRFinderPlus subclass equivalent}
\item{ab}{AMR package antibiotic code}
\item{drug_class}{The standardised class of the antimicrobial agent}
\item{drug_agent}{The name of the antimicrobial agent}
}
}
\source{
internal
}
\usage{
resfinder
}
\description{
A dataset containing the mapping between Resfinder drug agents and their standardised drug classes.
}
\keyword{datasets}
