% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rgi_drugs_table}
\alias{rgi_drugs_table}
\title{Table mapping CARD/RGI drug class and antibiotic columns}
\format{
\code{rgi_drugs_table} A data frame with 3 rows and 70 columns:
\itemize{
\item \code{RGI_DrugClassAgent}: RGI/CARD drug class / antibiotic that is not recognized by AMR pkg
\item \code{drug_class}: Valid drug class name recognised by AMR pkg
\item \code{drug_agent}: Valid antibiotic name recognised by AMR pkg
}
}
\usage{
rgi_drugs_table
}
\description{
Table mapping CARD/RGI drug class / antibiotic columns to class / antibiotic names recognised by AMR pkg
}
\keyword{datasets}
