% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rgi_EuSCAPE_raw}
\alias{rgi_EuSCAPE_raw}
\title{Example Resistance Gene Identifier (RGI) v6.0.6 Genotype Data from EuSCAPE project}
\format{
\code{rgi_EuSCAPE_raw} A data frame with 59,403 rows and 25 columns:
\itemize{
\item \code{ORF_ID}: Sample identifier
\item ...: RGI results columns
}
}
\source{
ENA BioProject \href{https://www.ebi.ac.uk/ena/browser/view/PRJEB10018}{PRJEB10018}.
See David \emph{et al.} (2019) \url{https://doi.org/10.1038/s41564-019-0492-8}.
}
\usage{
rgi_EuSCAPE_raw
}
\description{
Raw RGI v6.0.6 results file for Klebsiella pneumoniae from EuSCAPE project, one AMR determinant per row.
Includes only Perfect and Strict hits
Columns \code{Predicted_DNA}, \code{Predicted_Protein}, and \code{CARD_Protein_Sequence} were removed to reduce file size.
}
\keyword{datasets}
