% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rgi_raw}
\alias{rgi_raw}
\title{Example Resistance Gene Identifier (RGI) v6.0.6 Genotype Data}
\format{
\code{rgi_raw} A data frame with 21,203 rows and 28 columns:
\itemize{
\item \code{ORF_ID}: Sample identifier
\item ...: RGI results columns
}
}
\source{
Four colistin-resistant isolates from Bioproject PRJNA966919 \url{https://www.ncbi.nlm.nih.gov/datasets/genome/?bioproject=PRJNA966919}.

One genome with rRNA gene variant models detected (GCF_000249055.1).
}
\usage{
rgi_raw
}
\description{
Raw RGI v6.0.6 results file (run with \code{--include_loose}) for 12 genomes of multiple species, one AMR determinant per row.
Includes multiple species to cover all four model types currently detected by RGI (protein homolog model, protein variant model, protein overexpression model, and rRNA gene variant model)
Includes Perfect, Strict, and Loose hits to test \code{exclude_loose} parameter
}
\keyword{datasets}
