% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rgi_short_name_table}
\alias{rgi_short_name_table}
\title{Table mapping CARD/RGI Model ID and CARD Short Name}
\format{
\code{rgi_short_name_table} A data frame with 6445 rows and 2 columns:
\itemize{
\item \verb{Model ID}: RGI/CARD model ID column
\item \verb{CARD Short Name}: Shortened AMR determinant names
}
}
\source{
Two columns extracted from: \url{https://card.mcmaster.ca/latest/data} in aro_index.tsv
}
\usage{
rgi_short_name_table
}
\description{
Table mapping all CARD/RGI Model ID and CARD Short Name to convert long names into shortened AMR determinant names
}
\keyword{datasets}
