% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salm_raw}
\alias{salm_raw}
\title{Example Salmonella Genotype-Phenotype Data}
\format{
\code{salm_raw} A data frame with 115 rows and 7 columns:
\itemize{
\item \code{Sample}: Sample identifier
\item \code{Source}, \code{Serovar}: Non-AMR related information about each isolate
\item \code{CpL_Genotype}: List of genotypic parkers (separated by \verb{;})
\item \code{Ciprofloxacin}, \code{Levofloxacin}, \code{Moxifloxacin}: MIC value for each drug
}
}
\usage{
salm_raw
}
\description{
Raw genotype-phenotype data for \emph{Salmonella enterica} genomes, one row per sample.
}
\keyword{datasets}
