% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solo_ppv.R
\name{solo_ppv_analysis}
\alias{solo_ppv_analysis}
\title{Perform Solo PPV Analysis for AMR Markers}
\usage{
solo_ppv_analysis(
  geno_table,
  pheno_table,
  antibiotic,
  drug_class_list,
  geno_sample_col = NULL,
  pheno_sample_col = NULL,
  sir_col = NULL,
  plot_cols = c(R = "IndianRed", I = "orange", S = "lightgrey", NWT = "navy")
)
}
\arguments{
\item{geno_table}{A data frame containing genotype data, including at least one column labeled
\code{drug_class} for drug class information and one column for sample identifiers (specified via
\code{geno_sample_col} otherwise it is assumed the first column contains identifiers).}

\item{pheno_table}{A data frame containing phenotype data, which must include a column \code{drug_agent}
(with the antibiotic information) and a column with the resistance interpretation (S/I/R, colname
specified via \code{sir_col}).}

\item{antibiotic}{A character string specifying the antibiotic of interest to filter phenotype data.
The value must match one of the entries in the \code{drug_agent} column of \code{pheno_table}.}

\item{drug_class_list}{A character vector of drug classes to filter genotype data for markers
related to the specified antibiotic. Markers in \code{geno_table} will be filtered based on whether
their \code{drug_class} matches any value in this list.}

\item{geno_sample_col}{A character string (optional) specifying the column name in \code{geno_table}
containing sample identifiers. Defaults to \code{NULL}, in which case it is assumed the first column
contains identifiers.}

\item{pheno_sample_col}{A character string (optional) specifying the column name in \code{pheno_table}
containing sample identifiers. Defaults to \code{NULL}, in which case it is assumed the first column
contains identifiers.}

\item{sir_col}{A character string specifying the column name in \code{pheno_table} that contains
the resistance interpretation (SIR) data. The values should be interpretable as "R" (resistant),
"I" (intermediate), or "S" (susceptible).}

\item{plot_cols}{A named vector of colors for the plot. The names should be the phenotype categories
(e.g., "R", "I", "S", "NWT"), and the values should be valid color names or hexadecimal color codes.
Default colors are provided for resistant ("R"), intermediate ("I"), susceptible ("S"), and non-wild-type ("NWT").}
}
\value{
A list containing the following elements:
\describe{
\item{solo_stats}{A dataframe summarizing the PPV for resistance (R vs S/I) and NWT (R/I vs S),
including the number of positive hits, sample size, PPV, and 95\% confidence intervals for each marker.}
\item{combined_plot}{A combined ggplot object showing the PPV plot for the solo markers, and a bar plot
for the phenotype distribution.}
\item{solo_binary}{A dataframe with binary values indicating the presence or absence of the solo markers.}
\item{amr_binary}{A dataframe with binary values for the AMR markers, based on the input genotype and phenotype data.}
}
}
\description{
This function performs a Positive Predictive Value (PPV) analysis for AMR markers
associated with a given antibiotic and drug class. The function calculates the PPV
for solo markers (those with only one genetic marker relevant to the drug class)
and visualizes the results using various plots. It returns a list containing summary
statistics for each solo marker, and associated plots showing the breakdown of
resistance phenotypes, and PPV (with 95\% confidence interval) for each marker.
}
\examples{
\dontrun{
geno_table <- parse_amrfp("testdata/Ecoli_AMRfinderplus.tsv.gz", "Name")
pheno_table <- import_ncbi_ast("testdata/Ecoli_AST_NCBI.tsv.gz")

soloPPV_cipro <- solo_ppv_analysis(
  geno_table = geno_table,
  pheno_table = pheno_table,
  antibiotic = "Ciprofloxacin",
  drug_class_list = c("Quinolones"),
  sir_col = "Resistance phenotype"
)

soloPPV_mero <- solo_ppv_analysis(
  geno_table = geno_table,
  pheno_table = pheno_table,
  antibiotic = "Meropenem",
  drug_class_list = c("Carbapenems", "Cephalosporins"),
  sir_col = "Resistance phenotype"
)

# View the results
soloPPV_cipro$solo_stats
soloPPV_cipro$combined_plot
soloPPV_mero$solo_stats
soloPPV_mero$combined_plot
}
}
