% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{staph_ast_ebi}
\alias{staph_ast_ebi}
\title{S. aureus Example of Imported EBI Phenotype Data}
\format{
\code{staph_ast_ebi} A data frame with 218 rows and 46 columns representing all Staphylococcus phenotyping results for amikacin and doxycycline downloaded from EBI using \link{download_ebi}, and imported into AMRgen format using \link{import_ast}.

Columns include:
\itemize{
\item \code{id}: Sample identifier.
\item \code{drug_agent}: Antibiotic identifier, as class 'ab'.
\item \code{mic}: MIC data, as class 'mic'.
\item \code{disk}: Disk diffusion zone diameter data, as class 'disk'.
\item \code{pheno_provided}, \code{pheno_eucast}, \code{pheno_clsi}, \code{ecoff}: S/I/R phenotypes as downloaded from EBI, and as re-interpreted from mic/disk measures against EUCAST 2024 breakpoints.
\item ...: Additional data columns from EBI.
}
}
\source{
\url{https://www.ebi.ac.uk/amr}
}
\usage{
staph_ast_ebi
}
\description{
Phenotypes sourced from EBI AMR Portal using the \link{download_ebi} function and imported to AMRgen phenotype table format.
}
\keyword{datasets}
