% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{staph_ast_ncbi}
\alias{staph_ast_ncbi}
\title{S. aureus Example of Imported NCBI Phenotype Data}
\format{
\code{staph_ast_ncbi} A data frame with 143 rows and 19 columns representing all Staphylococcus aureus phenotyping results for amikacin and doxycycline downloaded from NCBI using \link{download_ncbi_ast}, imported into AMRgen format using \link{import_ast}.

Columns include:
\itemize{
\item \code{id}: Sample identifier.
\item \code{drug_agent}: Antibiotic identifier, as class 'ab'.
\item \code{mic}: MIC data, as class 'mic'.
\item \code{disk}: Disk diffusion zone diameter data, as class 'disk'.
\item \code{pheno_provided}, \code{pheno_eucast}: S/I/R phenotypes as downloaded from NCBI, and as re-interpreted from mic/disk measures against EUCAST 2024 breakpoints.
\item ...: Additional data columns from NCBI.
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pathogens/ast}
}
\usage{
staph_ast_ncbi
}
\description{
Phenotypes sourced from NCBI Biosamples using the \link{download_ncbi_ast} function and imported to AMRgen phenotype table format.
}
\keyword{datasets}
