% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{staph_ast_ncbi_cloud_raw}
\alias{staph_ast_ncbi_cloud_raw}
\title{S. aureus Example of Raw Phenotype Data Downloaded from NCBI via Google Cloud BigQuery}
\format{
\code{staph_ast_ncbi_cloud_raw} A data frame with 142 rows and 11 columns representing all Staphylococcus aureus phenotyping results for amikacin and doxycycline.

Columns include:
\itemize{
\item \code{BioSample}: Sample identifier.
\item \code{Antibiotic}: Antibiotic name.
\item \verb{Resistance phenotype}: S/I/R phenotypes as downloaded from NCBI.
\item ...: Additional data columns from NCBI.
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pathogens/ast}
}
\usage{
staph_ast_ncbi_cloud_raw
}
\description{
Phenotypes sourced from NCBI via \link{query_ncbi_bq_ast} function, without reformating.
}
\keyword{datasets}
