% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{staph_geno_ebi}
\alias{staph_geno_ebi}
\title{S. aureus Example of Imported EBI Genotype Data}
\format{
\code{staph_geno_ebi} A data frame with 198344 rows and 34 columns representing all Staphylococcus genotyping results downloaded from EBI using \link{download_ebi}, and imported into AMRgen format using \link{import_amrfp}.

Columns include:
\itemize{
\item \code{id}: Sample identifier.
\item \code{drug_agent}, \code{drug_class}: Antibiotic agent and class, determined by parsing AMRfinderplus \code{subclass} field in the downloaded file.
\item \code{gene}, \code{node}, \code{marker}: gene symbol, parsed from \code{amr_element_symbol} field in the downloaded file.
\item \code{mutation}: mutation within gene, parsed into HGVS nomenclature format from \code{amr_element_symbol} field in the downloaded file.
\item \code{marker.label}: label for genotype marker, combining \code{gene} and \code{mutation} information (deletion variants represented as \code{"gene:-"}).
\item ...: Additional data columns from EBI.
}
}
\source{
\url{https://www.ebi.ac.uk/amr}
}
\usage{
staph_geno_ebi
}
\description{
Phenotypes sourced from EBI AMR Portal using the \link{download_ebi} function and imported to AMRgen phenotype table format.
}
\keyword{datasets}
