% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{staph_geno_ncbi_cloud_raw}
\alias{staph_geno_ncbi_cloud_raw}
\title{S. aureus Example of Raw Genotype Data Downloaded from NCBI via Google Cloud BigQuery}
\format{
\code{staph_geno_ncbi_cloud_raw} A data frame with 4064 rows and 9 columns representing all Staphylococcus aureus genotyping results for markers associated with class aminoglycoside or tetracycline.

Columns include:
\itemize{
\item \code{biosample_acc}: Sample identifier.
\item \code{scientific_name}: Organism name.
\item \verb{Gene symbol}, \code{Class}, \code{Subclass}, \verb{Element type}, \verb{Element subtype}, \code{Method}, \code{Hierarchy_node}: Key results fields from AMRfinderplus.
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pathogens/microbigge/}
}
\usage{
staph_geno_ncbi_cloud_raw
}
\description{
AMRfinderplus genotypes sourced from NCBI via \link{query_ncbi_bq_ast} function, without reformating.
}
\keyword{datasets}
