% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amr_upset.R
\name{stripRangeValues}
\alias{stripRangeValues}
\title{Strip Range Values from MIC Object}
\usage{
stripRangeValues(x)
}
\arguments{
\item{x}{A vector, such as MIC or disk diffusion values.}
}
\value{
If the input was a \code{mic} object (from the AMR package), any values expressed as ranges
will be removed before returning (ie if all are ranges, the returned object will be empty).
If not a \code{mic} object, the input vector will be returned unmodified.
}
\description{
This function removes any range values (i.e., values containing \code{<}, \code{>}, or \code{=}) from an
object of class \code{mic} (from the \code{AMR} package). It only modifies the object if it is of class \code{mic}.
}
\examples{
# Example usage
x <- AMR::as.mic(c("0.5", "0.25", "1", "0.5", "<0.5", ">=0.25", ">0.25"))
stripRangeValues(x)

x <- AMR::as.disk(c(32,35,31,16,12))
stripRangeValues(x)

}
