% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_geno}
\alias{summarise_geno}
\title{Summarise a Genotype Table}
\usage{
summarise_geno(
  geno_table,
  sample_col = "Name",
  marker_col = "marker",
  drug_col = "drug_agent",
  class_col = "drug_class",
  gene_col = "gene",
  variation_col = "variation type",
  force_ab = FALSE
)
}
\arguments{
\item{geno_table}{A tibble or data frame containing genotype data, in the format output by \link{import_amrfp}.}

\item{sample_col}{Character. Name of the column containing sample identifiers. Default is \code{"Name"}.}

\item{marker_col}{Character. Name of the column containing marker identifiers. Default is \code{"marker"}.}

\item{drug_col}{Character. Name of the column containing drug agent identifiers. Default is \code{"drug_agent"}. If this is of class 'ab' the entries will be annotated with their full antibiotic names, converted using \link[AMR:as.ab]{AMR::as.ab}. If this is desired behaviour but the class is not 'ab', set \code{force_ab=TRUE}.}

\item{class_col}{Character. Name of the column containing drug class identifiers. Default is \code{"drug_class"}.}

\item{gene_col}{Character. Name of the column containing gene identifiers. Default is \code{"gene"}.}

\item{variation_col}{Character. Name of the column containing variation type identifiers. Default is \code{"variation type"}.}

\item{force_ab}{Logical. If \code{TRUE}, attempts to convert entries in \code{drug_col} to antibiotic names using \link[AMR:as.ab]{AMR::as.ab} even if this column is not of class \code{"ab"} Default is \code{FALSE}.}
}
\value{
A named list with the following elements:
\describe{
\item{uniques}{A tibble of the number of unique samples, markers, genes, drugs, classes and variation types detected in \code{geno_table}.}
\item{per_type}{A tibble of unique counts of samples, markers, genes, drugs, and classes per variation type.}
\item{drugs}{A tibble listing the drugs and/or drug classes represented in the table, and the associated number of unique markers, unique samples, and total hits for each drug/class.}
\item{markers}{A tibble listing the markers represented in the table, and the associated drugs/classes and variation types (if present). Number indicates the count of hits detected per marker.}
}
}
\description{
\code{summarise_geno()} computes summary information for a genotype table.
}
\details{
The function automatically adapts to the presence or absence of columns in \code{geno_table}.
The \code{force_ab} parameter allows the addition of full antibiotic names using the \code{ab_name()} function even when the first column is not recognized as an \code{"ab"} object.
}
\examples{

geno_table <- import_amrfp(ecoli_geno_raw)
summarise_geno(geno_table)

}
