% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_geno_pheno}
\alias{summarise_geno_pheno}
\title{Summarise the intersection of a genotype table and a phenotype table}
\usage{
summarise_geno_pheno(
  geno_table,
  pheno_table,
  geno_sample_col = NULL,
  pheno_sample_col = NULL,
  pheno_cols,
  drug_col = "drug_agent",
  class_col = "drug_class",
  marker_col = "marker",
  gene_col = "gene",
  variation_col = "variation type",
  mic_col = "mic",
  disk_col = "disk",
  spp_col = "spp_pheno",
  method_cols = c("method", "platform", "guideline", "source")
)
}
\arguments{
\item{geno_table}{A tibble or data frame containing genotype data, in the format output by \link{import_amrfp}.}

\item{pheno_table}{A tibble or data frame containing phenotype data, in the format output by \link{import_ast}.}

\item{geno_sample_col}{Character. Name of the column in the genotype table containing sample identifiers. Default is \code{"Name"}.}

\item{pheno_sample_col}{Character. Name of the column in the phenotype table containing sample identifiers. Default is \code{"id"}.}

\item{pheno_cols}{Vector. Vector giving names of columns in the phenotype table containing categorical phenotype calls (S/I/R or NWT/WT). Default is any columns beginning with \code{"pheno"} or \code{"ecoff"}.}

\item{drug_col}{Character. Name of the column in the phenotype table containing drug agent identifiers. Default is \code{"drug_agent"}. Entries will be annotated with their full antibiotic names, converted using \code{\link[AMR:as.ab]{AMR::as.ab()}}. If the genotype table contains a column indicating individual agents it should share this same name.}

\item{class_col}{Character. Name of the column in the genotype table containing drug class identifiers. Default is \code{"drug_class"}. This should be antibiotic group names understood by the AMR pkg, as per \code{drug_col} columns generated by the \code{\link[=import_geno]{import_geno()}} functions.}

\item{marker_col}{Character. Name of the column in the genotype table containing marker identifiers. Default is \code{"marker"}.}

\item{gene_col}{Character. Name of the column in the genotype table containing gene identifiers. Default is \code{"gene"}.}

\item{variation_col}{Character. Name of the column in the genotype table containing variation type identifiers. Default is \code{"variation type"}.}

\item{mic_col}{Character. Name of the column in the phenotype table containing MIC measurements Default is \code{"mic"}.}

\item{disk_col}{Character. Name of the column in the phenotype table containing disk diffusion zone measurements. Default is \code{"disk"}.}

\item{spp_col}{Character. Name of the column in the phenotype table containing species names. Default is \code{"spp_pheno"}.}

\item{method_cols}{Vector. Vector giving names of columns in the phenotype table containing method or source information by which to summarise MIC/disk data. Default is \code{c("method", "platform", "guideline", "source")}.}
}
\value{
A named list with the following elements:
\itemize{
\item \code{drugs_with_pheno}: A tibble listing the drugs included in the table, and the associated number of samples with MIC measures, disk measures, neither or both, for each drug; restricted to samples that also appear in the genotype table.
\item \code{geno_hits}: A tibble listing the drugs and/or drug classes corresponding to drugs with phenotypes, and the associated number of unique markers, unique samples, and total hits for each drug/class amongst samples with corresponding phenotype data.
\item \code{geno_markers}: A tibble listing the genotypic markers in the genotype table corresponding to drugs with phenotypes, and the associated drugs/classes and variation types (if present). Number indicates the count of hits detected per marker, amongst samples with corresponding phenotype data.
\item \code{pheno_counts_list}: A list of tibbles, each corresponding to a unique categorical phenotype column in the input, indicating the counts of each phenotypic category per drug and species; restricted to samples that also appear in the genotype table.
}
}
\description{
Compares a genotype table and phenotype table, to summarise the number of samples present in both tables. For each drug in the phenotype table, summarise the number of phenotype observations, and details of genotype markers detected for the associated drug class/es.
}
\details{
The function automatically adapts to the presence or absence of columns in \code{pheno_table}.
The \code{force_ab} parameter allows the addition of full antibiotic names using the \code{ab_name()} function even when the first column is not recognized as an \code{"ab"} object.
}
\examples{
staph_geno_pheno <- summarise_geno_pheno(staph_geno_ebi, staph_ast_ebi,
  pheno_cols = c("pheno_clsi", "pheno_provided")
)
staph_geno_pheno
}
