% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_pheno}
\alias{summarise_pheno}
\title{Summarise a Phenotype Table}
\usage{
summarise_pheno(
  pheno_table,
  sample_col = "id",
  drug_col = "drug_agent",
  mic_col = "mic",
  disk_col = "disk",
  spp_col = "spp_pheno",
  pheno_cols = NULL,
  method_cols = c("method", "platform", "guideline", "source"),
  force_ab = FALSE
)
}
\arguments{
\item{pheno_table}{A tibble or data frame containing phenotype data, in the format output by \link{import_ast}.}

\item{sample_col}{Character. Name of the column containing sample identifiers. Default is \code{"id"}.}

\item{drug_col}{Character. Name of the column containing drug agent identifiers. Default is \code{"drug_agent"}. If this is of class 'ab' the entries will be annotated with their full antibiotic names, converted using \link[AMR:as.ab]{AMR::as.ab}. If this is desired behaviour but the class is not 'ab', set \code{force_ab=TRUE}.}

\item{mic_col}{Character. Name of the column containing MIC data. Default is \code{"mic"}.}

\item{disk_col}{Character. Name of the column containing drug class identifiers. Default is \code{"disk"}.}

\item{spp_col}{Character. Name of the column containing species names Default is \code{"spp_pheno"}.}

\item{pheno_cols}{Vector. Vector giving names of columns containing categorical phenotype calls (S/I/R or NWT/WT). Default is any columns beginning with \code{"pheno"} or \code{"ecoff"}.}

\item{method_cols}{Vector. Vector giving names of columns containing method or source information by which to summarise MIC/disk data. Default is \code{c("method", "platform", "guideline", "source")}.}

\item{force_ab}{Logical. If \code{TRUE}, attempts to convert entries in \code{drug_col} to antibiotic names using \link[AMR:as.ab]{AMR::as.ab} even if this column is not of class \code{"ab"} Default is \code{FALSE}.}
}
\value{
A named list with the following elements:
\describe{
\item{uniques}{A tibble of the number of unique samples, drugs, organisms, and methods detected in \code{pheno_table}.}
\item{drugs}{A tibble listing the drugs included in the table, and the associated number of samples with MIC measures, disk measures, neither or both, for each drug and species.}
\item{details}{A tibble listing more details of the methods of assay measurements, per drug and species.}
\item{pheno_counts_list}{A list of tibbles, each corresponding to a unique categorical phenotype column in the input, indicating the counts of each phenotypic category per drug and species.}
}
}
\description{
\code{summarise_pheno()} computes summary information for a phenotype table.
}
\details{
The function automatically adapts to the presence or absence of columns in \code{pheno_table}.
The \code{force_ab} parameter allows the addition of full antibiotic names using the \code{ab_name()} function even when the first column is not recognized as an \code{"ab"} object.
}
\examples{

summarise_pheno(staph_ast_ebi)

summarise_pheno(staph_ast_ebi, pheno_cols = c("pheno_provided", "pheno_clsi", "ecoff"))

}
