% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{amr_logistic}
\alias{amr_logistic}
\title{AMR Logistic Regression Analysis}
\usage{
amr_logistic(
  geno_table,
  pheno_table,
  antibiotic,
  drug_class_list,
  geno_sample_col = NULL,
  pheno_sample_col = NULL,
  sir_col = "pheno",
  ecoff_col = "ecoff",
  maf = 10,
  fit_glm = FALSE,
  single_plot = TRUE,
  colors = c("maroon", "blue4"),
  axis_label_size = 9,
  marker_col = "marker.label"
)
}
\arguments{
\item{geno_table}{A data frame containing the genotype data.}

\item{pheno_table}{A data frame containing the phenotypic data.}

\item{antibiotic}{A character string specifying the antibiotic to model using logistic regression.}

\item{drug_class_list}{A vector of drug class names. Used to subset the relevant markers for analysis.}

\item{geno_sample_col}{(Optional) A character string specifying the column in \code{geno_table} that identifies the sample IDs. Defaults to \code{NULL}.}

\item{pheno_sample_col}{(Optional) A character string specifying the column in \code{pheno_table} that identifies the sample IDs. Defaults to \code{NULL}.}

\item{sir_col}{(Optional) A character string specifying the column in \code{pheno_table} that contains the phenotype values (e.g., resistance/susceptibility). Defaults to \code{"pheno"}.}

\item{ecoff_col}{(Optional) A character string specifying the column in \code{pheno_table} containing the ECOFF (epidemiological cutoff) values. Defaults to \code{"ecoff"}.}

\item{maf}{(Optional) An integer specifying the minimum allele frequency (MAF) threshold. Markers with a MAF lower than this value will be excluded. Defaults to 10.}

\item{fit_glm}{(Optional) Change to TRUE to fit model with glm. Otherwise fit model with logistf (default).}

\item{single_plot}{(Optional) A logical value. If \code{TRUE}, a single plot is produced comparing the estimates for resistance (\code{R}) and non-resistance (\code{NWT}). Otherwise, two plots are printed side-by-side. Defaults to \code{TRUE}.}

\item{colors}{(Optional) A vector of two colors, to use for R and NWT models in the plots. Defaults to \code{c("maroon", "blue4")}.}

\item{axis_label_size}{(Optional) A numeric value controlling the size of axis labels in the plot. Defaults to 9.}

\item{marker_col}{(Optional) Name of the column containing the marker identifiers, whose unique values will be treate as predictors in the regression. Defaults to \code{"marker"}.}
}
\value{
A list with the following components:
\itemize{
\item \code{bin_mat}: The binary matrix of genetic data and phenotypic resistance information.
\item \code{modelR}: The fitted logistic regression model for resistance (\code{R}).
\item \code{modelNWT}: The fitted logistic regression model for non-resistance (\code{NWT}).
\item \code{plot}: A ggplot object comparing the estimates for resistance and non-resistance with corresponding statistical significance indicators.
}
}
\description{
Performs logistic regression to analyze the relationship between genetic markers and phenotype (R, and NWT) for a specified antibiotic.
}
\examples{
# Example usage of the amr_logistic function
result <- amr_logistic(
  geno_table = import_amrfp(ecoli_geno_raw, "Name"),
  pheno_table = ecoli_ast,
  sir_col = "pheno_clsi",
  antibiotic = "Ciprofloxacin",
  drug_class_list = c("Quinolones"),
  maf = 10
)
# To access the plot:
print(result$plot)
}
