% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMRgen-package.R
\docType{package}
\name{AMRgen-package}
\alias{AMRgen}
\alias{AMRgen-package}
\title{AMRgen: Genotype-Phenotype Integration for Antimicrobial Resistance Research}
\description{
The AMRgen package provides tools for interpreting antimicrobial resistance (AMR) genes, integrating genomic data with phenotypic antimicrobial susceptibility testing (AST) data, and calculating genotype-phenotype associations. This package extends the functionality of the AMR package, enabling researchers to fetch MIC distribution data from EUCAST, visualise multidrug resistance patterns using UpSet plots, and generate NCBI-compliant antibiograms. Designed for AMR research and surveillance, AMRgen aims to support data-driven decision-making in combating AMR on a global scale.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://AMRverse.github.io/AMRgen/}
  \item \url{https://github.com/AMRverse/AMRgen}
}

}
\author{
\strong{Maintainer}: Matthijs S. Berends \email{m.s.berends@umcg.nl} (\href{https://orcid.org/0000-0001-7620-1800}{ORCID})

Authors:
\itemize{
  \item Natacha Couto (\href{https://orcid.org/0000-0002-9152-5464}{ORCID})
  \item Zoe A. Dyson \email{zoe.dyson@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-8887-3492}{ORCID})
  \item Jane Hawkey (\href{https://orcid.org/0000-0001-9661-5293}{ORCID})
  \item Kat Holt (\href{https://orcid.org/0000-0003-3949-2471}{ORCID})
}

}
\keyword{internal}
