% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_mic_gen.R
\name{assay_by_var}
\alias{assay_by_var}
\title{Generate a Stacked Bar Plot of MIC Values Colored by Gene Symbol for Each Antibiotic}
\usage{
assay_by_var(
  pheno_table,
  antibiotic,
  measure = "mic",
  facet_var = NULL,
  species = NULL,
  bp_site = NULL,
  bp_S = NULL,
  bp_R = NULL,
  ecoff = NULL,
  marker_free_strains = NULL,
  colour_by = NULL,
  cols = NULL
)
}
\arguments{
\item{pheno_table}{Phenotype table in standard format as per import_ast().}

\item{antibiotic}{Name of the antibiotic.}

\item{measure}{Field name containing the assay measurements to plot (default "mic").}

\item{facet_var}{(optional) Field name containing a field to facet on (default NULL).}

\item{species}{(optional) Name of species, so we can retrieve breakpoints to print at the top of the plot to help interpret it.}

\item{bp_site}{(optional) Breakpoint site to retrieve (only relevant if also supplying \code{species} to retrieve breakpoints, and not supplying breakpoints via \code{bp_S}, \code{bp_R}, \code{ecoff}).}

\item{bp_S}{(optional) S breakpoint}

\item{bp_R}{(optional) R breakpoint}

\item{ecoff}{(optional) ECOFF breakpoint}

\item{marker_free_strains}{(optional) Vector of sample names to select to get their own plot. Most useful for defining the set of strains with no known markers associated with the given antibiotic, so you can view the distribution of assay values for strains expected to be wildtype, which can help to identify issues with the assay.}

\item{colour_by}{(optional) Field name containing a field to colour bars by (default NULL, which will colour each bar to indicate whether the value is expressed as a range or not)}

\item{cols}{(optional) Manual colour scale to use for plot. If NULL, \code{colour_by} variable is of class 'sir', bars will by default be coloured using standard SIR colours.}
}
\value{
A list containing
\item{plot}{Main plot with all samples that have assay data for the given antibiotic}
\item{plot_nomarkers}{Additional plot showing only those samples listed in \code{marker_free_strains}}
}
\description{
This function creates a stacked bar plot using \code{ggplot2}, where the x-axis represents MIC (Minimum Inhibitory Concentration) or disk values, the y-axis represents their frequency, and the bars are colored to indicate whether the assay value is expressed as a range or not. It can optionally be faceted on an additional categorical variable.
}
