% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccuring_ppv.R
\name{cooccuring_ppv_analysis}
\alias{cooccuring_ppv_analysis}
\title{Co-occuring PPV Analysis}
\usage{
cooccuring_ppv_analysis(
  geno_table,
  pheno_table,
  antibiotic,
  drug_class_list,
  geno_sample_col = NULL,
  pheno_sample_col = NULL,
  sir_col = NULL,
  marker_col = "marker",
  keep_assay_values = TRUE,
  min = 1,
  axis_label_size = 9,
  pd = position_dodge(width = 0.8),
  plot_cols = c(R = "IndianRed", NWT = "navy")
)
}
\arguments{
\item{geno_table}{A data frame containing genotype data.}

\item{pheno_table}{A data frame containing phenotype data.}

\item{antibiotic}{String. The specific antibiotic to analyze (e.g., "Ceftazidime").}

\item{drug_class_list}{Vector of strings. The drug classes to filter for (e.g., c("Cephalosporins")).}

\item{geno_sample_col}{String. Column name for sample IDs in genotype table.}

\item{pheno_sample_col}{String. Column name for sample IDs in phenotype table.}

\item{sir_col}{String. Column name containing S/I/R data in the phenotype table.}

\item{marker_col}{String. Default "marker". Name of the column identifying genes.}

\item{keep_assay_values}{Logical. Default TRUE. Whether to keep MIC/Disk columns.}

\item{min}{Integer. Default 1. Minimum number of isolates required for a profile to be plotted.}

\item{axis_label_size}{Integer. Font size for axis text.}

\item{pd}{Position dodge object for plotting.}

\item{plot_cols}{Named vector of colors for "R" and "NWT".}
}
\value{
A list containing:
\item{stats}{Data frame of calculated PPV statistics per profile.}
\item{plot}{The combined ggplot/patchwork object.}
\item{data}{The filtered data frame used for the analysis.}
}
\description{
This function performs a Positive Predictive Value (PPV) analysis for single or multiple AMR markers associated with a given antibiotic and drug class. It calculates the PPV for single and co-occurring markers and visualizes the results using UpSet-style matrix plot.
}
\examples{
\dontrun{
geno_table <- import_amrfp(ecoli_geno_raw, "Name")
head(ecoli_ast)
cooccuringPPV_cipro <- cooccuring_ppv_analysis(
  geno_table = geno_table,
  pheno_table = ecoli_ast,
  antibiotic = "Ciprofloxacin",
  drug_class_list = c("Quinolones"),
  sir_col = "pheno_provided"
)
}
}
