% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ebi.R
\name{download_ebi_geno}
\alias{download_ebi_geno}
\title{Download antimicrobial genotype data from the EBI AMR Portal}
\usage{
download_ebi_geno(
  user_genus = NULL,
  user_subclass = NULL,
  user_class = NULL,
  user_antibiotic_name = NULL,
  user_release = NULL
)
}
\arguments{
\item{user_genus}{String specifying a bacterial genus to return data for (default NULL, will pull all taxa)}

\item{user_subclass}{String specifying an antibiotic subclass to filter on (default NULL, check NCBI Subclass for valid terms).}

\item{user_class}{String specifying an antibiotic subclass to filter on (default NULL, check NCBI Class for valid terms).}

\item{user_antibiotic_name}{String specifying an antibiotic to return data for (default NULL, will pull all antibiotics).}

\item{user_release}{String specifying the data release to download (default NULL, will pull latest release).}
}
\value{
A data frame containing EBI genotype data
}
\description{
This function will retrieve genotype data from the EBI AMR Portal, https://www.ebi.ac.uk/amr. The portal uses AMRfinderplus to identify AMR-associated genotypes, but the results are processed and not all fields returned by AMRfinderplus are included. See https://www.ebi.ac.uk/amr/about/#AMR-Genotypes for more information, and https://github.com/ncbi/amr/wiki/class-subclass for valid class and subclass terms.
}
\examples{
\dontrun{
ebi_genotypes <- import_ebi()

amp_sal_ebi <- import_ebi(
    user_genus="Salmonella",
    user_release="2025-12",
    user_subclass="beta-lactam"
)
}
}
\keyword{internal}
