% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli_ast}
\alias{ecoli_ast}
\title{E. coli NCBI AST Example Data, Re-interpreted with AMR Package}
\format{
\subsection{\code{ecoli_ast} A data frame with 4170 rows and 10 columns representing reinterpreted data from the NCBI AST browser.}{

Columns include:
\itemize{
\item \code{id}: Sample identifier, imported from the \verb{#BioSample} column in the raw input.
\item \code{spp_pheno}: Species identifier, interpreted from \verb{Scientific name} using \code{as.mo}, used to interpret \code{ecoff} and \code{pheno} columns.
\item \code{drug_agent}: Antibiotic code, interpreted from \code{Antibiotic} using \code{as.ab}, used to interpret \code{ecoff} and \code{pheno} columns.
\item \code{ecoff}: WT/NWT classification, interpreted using \code{as.sir}.
\item \code{pheno}: S/I/R classification, interpreted using \code{as.sir}.
\item \code{mic}: Minimum inhibitory concentration, formatted using \code{as.mic}, used to interpret \code{ecoff} and \code{pheno} columns.
\item \code{disk}: Disk diffusion zone, formatted using \code{as.disk}, used to interpret \code{ecoff} and \code{pheno} columns.
\item \code{guideline}: Interpretation guidelines used to interpret \code{ecoff} and \code{pheno} columns.
\item \verb{Scientific name}: Species identifier, from input file.
\item \code{Antibiotic}: Antibiotic name, from input file.
\item \verb{Testing standard}: Interpretation standard (EUCAST or CLSI), from input file.
\item \verb{Measurement sign}: Measurement sign (>, <, =, etc.) relating to MIC measurement, from input file.
\item \code{MIC (mg/L)}: Minimum inhibitory concentration, from input file.
\item \verb{Disk diffusion (mm)}: Disk diffusion zone, from input file.
\item \verb{Resistance phenotype}: Resistance call (SIR), from input file.
\item ...: Additional metadata columns from the NCBI AST export.
}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pathogens/ast}
}
\usage{
ecoli_ast
}
\description{
A subset of E. coli phenotype data from the NCBI AST browser.
}
\keyword{datasets}
