% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli_ast_raw}
\alias{ecoli_ast_raw}
\title{E. coli NCBI AST Example Data}
\format{
\subsection{\code{ecoli_ast_raw} A data frame with 10 rows and 17 columns representing unprocessed data from the NCBI AST browser.}{

Columns include:
\itemize{
\item \verb{#BioSample}: Sample identifier.
\item \verb{Scientific name}: Species identifier.
\item \code{Antibiotic}: Antibiotic name.
\item \verb{Testing standard}: Interpretation standard (EUCAST or CLSI).
\item \verb{Measurement sign}: Measurement sign (>, <, =, etc.) relating to MIC measurement.
\item \code{MIC (mg/L)}: Minimum inhibitory concentration.
\item \verb{Disk diffusion (mm)}: Disk diffusion zone.
\item \verb{Resistance phenotype}: Resistance call (SIR) as submitted.
\item ...: Additional metadata columns from the NCBI AST export.
}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pathogens/ast}
}
\usage{
ecoli_ast_raw
}
\description{
A subset of E. coli phenotype data from the NCBI AST browser.
}
\keyword{datasets}
