% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli_geno_raw}
\alias{ecoli_geno_raw}
\title{E. coli Genotype Example Data}
\format{
\subsection{\code{ecoli_geno_raw} A data frame with 45228 rows and 28 columns representing genotyping results from AMRFinderPlus.}{

Columns include:
\itemize{
\item \code{Name}: Sample identifier.
\item \verb{Gene symbol}: Gene symbol in NCBI RefGene.
\item \verb{Hierarchy node}: Node in NCBI hierarchy.
\item \code{Class}, \code{Subclass}: Drug class(es) associated with the marker (from NCBI RefGene).
\item \verb{\% Coverage of reference sequence}, \verb{\% Identity to reference sequence}, \verb{Accession of closest sequence}: Sequence match information.
\item ...: Additional metadata columns from the AMRFinderPlus output.
}
}
}
\source{
\url{https://github.com/ncbi/amr/wiki/Interpreting-results}
}
\usage{
ecoli_geno_raw
}
\description{
Genotypes called using AMRFinderPlus (v3.12.8, DB 2024-01-31.1), sourced from the AllTheBacteria project.
}
\keyword{datasets}
