% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_binary_matrix.R
\name{get_combo_matrix}
\alias{get_combo_matrix}
\title{Add marker combinations to a binary geno-pheno matrix}
\usage{
get_combo_matrix(binary_matrix, assay = NULL)
}
\arguments{
\item{binary_matrix}{A geno-pheno binary matrix, output by \code{get_binary_matrix}}

\item{assay}{(optional) Name of an assay column to filter on, so that the matrix returned only includes samples with assay data of this type available}
}
\description{
Given a geno-pheno binary marker matrix, output by \code{get_binary_matrix},
this function identifies marker combination, and adds combination ids,
and the count of markers detected per sample, as new columns.
}
\examples{
\dontrun{
combo_matrix <- get_combo_matrix(binary_matrix)
}

}
