% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_abricate.R
\name{import_abricate}
\alias{import_abricate}
\title{Import and Process Abricate Results}
\usage{
import_abricate(
  input_table,
  sample_col = "FILE",
  gene_col = "GENE",
  product_col = "PRODUCT",
  ab_col = "RESISTANCE",
  db = "resfinder"
)
}
\arguments{
\item{input_table}{A character string specifying a dataframe or path to the Abricate results table.}

\item{sample_col}{A character string specifying the column that identifies samples in the dataset (default \code{"FILE"}).}

\item{gene_col}{A character string specifying the column that identifies gene symbols in the dataset (default \code{"GENE"}).}

\item{product_col}{A character string specifying the column that identifies product names in the dataset (default \code{"PRODUCT"}).}

\item{ab_col}{A character string specifying the column that identifies which drug/s each detected gene is associated with (default \code{"RESISTANCE"}).}

\item{db}{A character string specifying which AMR gene database Abricate was run with (currently only \code{"resfinder"} is supported).}
}
\value{
A tibble containing the processed AMR elements, with harmonised gene names, mapped drug agents, and drug classes which can be used for other functions of the ARMgen package.
}
\description{
This function imports and processes Abricate results, extracting antimicrobial resistance (AMR) elements and mapping them to standardised antibiotic names and drug classes. Currently supports results generated using the ResFinder database.
}
\details{
The function performs the following steps:
\itemize{
\item Reads the Abricate output table via the internal \code{process_input} function.
\item Standardises the sample column name 'id'.
\item Assigns standardised column names for genes, markers, and sets variation type to "Gene presence detected".
\item Splits multiple resistance annotations (separated by semicolons) into separate rows.
\item Converts drug agent names to the \code{"ab"} class from the AMR package and maps these to classes compatible with the output of \code{\link[=import_amrfp]{import_amrfp()}} and \code{\link[=import_kleborate]{import_kleborate()}}.
}
}
\examples{
\dontrun{
geno_table <- import_abricate("path/to/abricate_resfinder.tsv")
}
}
