% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_amrfp.R
\name{import_amrfp_ebi_ftp}
\alias{import_amrfp_ebi_ftp}
\title{Import EBI-processed AMRFinderPlus Genotypes}
\usage{
import_amrfp_ebi_ftp(
  input_table,
  sample_col = "BioSample_ID",
  element_symbol_col = "amr_element_symbol",
  element_type_col = "element_type",
  element_subtype_col = "element_subtype",
  subclass_col = "subclass",
  class_col = "class",
  amrfp_drugs = amrfp_drugs_table
)
}
\arguments{
\item{input_table}{A character string specifying the path to the EBI genotype table (R object, or file path to a TSV file).}

\item{sample_col}{A character string specifying the column that identifies samples in the dataset (default "BioSample_ID").}

\item{element_symbol_col}{A character string specifying the column that contains the content of the AMRfinderplus 'Element symbol' field (default "amr_element_symbol").}

\item{element_type_col}{A character string specifying the column that contains the content of the AMRfinderplus 'Element type' field (default "element_type").}

\item{element_subtype_col}{A character string specifying the column that contains the content of the AMRfinderplus 'Element subtype' field (default "element_subtype").}

\item{subclass_col}{A character string specifying the column that contains the content of the AMRfinderplus 'Subclass' field (default "subclass").}

\item{class_col}{A character string specifying the column that contains the content of the AMRfinderplus 'Class' field (default "class").}

\item{amrfp_drugs}{Data from mapping drug names to classes (default amrfp_drugs_table).}
}
\value{
A tibble containing the processed AMR elements, with harmonised gene names, drug agents (mapped by ), and drug classes. The output retains the original columns from the AMRFinderPlus table along with the newly mapped variables.
}
\description{
This function imports processed EBI-processed AMRFinderPlus genotyping results. The expected input is genotype data retrieved from the \href{ftp://ftp.ebi.ac.uk/pub/databases/amr_portal/releases/}{EBI AMR Portal FTP site} either directly or via the function \code{\link[=download_ebi]{download_ebi()}}.
Note that files downloaded from the \href{https://www.ebi.ac.uk/amr/data/?view=predictions}{EBI AMR Portal web browser} are formatted differently.
}
\details{
These data are pre-processed by EBI to match NCBI class/subclass to CARD's antibiotic resistance ontology (ARO), however for consistency this function will re-process the data to generate 'drug_agent' and 'drug_class' fields consistent with the \code{\link[=import_amrfp]{import_amrfp()}} function (the EBI fields 'antibiotic*' are also retained).
Note several AMRfinderplus fields are excluded from EBI files, including hierarchy node, method, percent identity and coverage; therefore unlike the \code{\link[=import_amrfp]{import_amrfp()}} function, this function cannot assign 'variation type' or 'node'.

The function performs the following steps:
\itemize{
\item Reads the EBI-processed genotype table.
\item Maps AMRFinderPlus subclasses to standardised drug agent and drug class
names using \code{amrfp_drugs} (EBI-mappings are retained in ('antibiotic*' fields.)
\item Converts drug agent names to the \code{"ab"} class from the AMR package.
This processing ensures compatibility with downstream AMR analysis workflows.
}
}
\examples{
\dontrun{
# Download quinolone-related genotype data for E. coli, from EBI
ebi_geno_raw <- download_ebi(data="genotype", species = "Escherichia coli", 
                        geno_subclass="QUINOLONE")

# Format the file for import
ebi_geno <- import_amrfp_ebi_ftp(ebi_geno_raw)
}
}
