% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pheno.R
\name{import_ast}
\alias{import_ast}
\title{Import and Process AST Data from an EBI or NCBI antibiogram File}
\usage{
import_ast(
  input,
  format = "ebi",
  interpret_eucast = FALSE,
  interpret_clsi = FALSE,
  interpret_ecoff = FALSE,
  species = NULL,
  ab = NULL
)
}
\arguments{
\item{input}{A string representing a dataframe, or a path to an input file, containing the AST data in EBI antibiogram format. These files can be downloaded from the EBI AMR browser, e.g. https://www.ebi.ac.uk/amr/data/?view=experiments}

\item{format}{A string indicating the format of the data, either "ebi" (default) or "ncbi". This determines whether the data is passed on to the \code{import_ebi_ast} or \code{import_ncbi_ast}()` function to process.}

\item{interpret_eucast}{A logical value (default is FALSE). If \code{TRUE}, the function will interpret the susceptibility phenotype (SIR) for each row based on the MIC or disk diffusion values, against ECOFF human breakpoints. These will be reported in a new column \code{pheno_eucast}, of class 'sir'.}

\item{interpret_clsi}{A logical value (default is FALSE). If \code{TRUE}, the function will interpret the susceptibility phenotype (SIR) for each row based on the MIC or disk diffusion values, against CLSI human breakpoints. These will be reported in a new column \code{pheno_clsi}, of class 'sir'.}

\item{interpret_ecoff}{A logical value (default is FALSE). If \code{TRUE}, the function will interpret the wildtype vs nonwildtype status for each row based on the MIC or disk diffusion values, against epidemiological cut-off (ECOFF) values. These will be reported in a new column \code{ecoff}, of class 'sir' and coded as 'R' (nonwildtype) or 'S' (wildtype).}

\item{species}{(optional) Name of the species to use for phenotype interpretation. By default, the organism field in the input file will be assumed to specify the species for each sample, but if this is missing or you want to override it in the interpretation step, you may provide a single species name via this parameter.}

\item{ab}{(optional) Name of the antibiotic to use for phenotype interpretation. By default, the antibiotic field in the input file will be assumed to specify the antibiotic for each sample, but if this is missing or you want to override it in the interpretation step, you may provide a single antibiotic name via this parameter.}

\item{source}{(optional) A single value to record as the source of these data points, e.g. "EBI_browser". By default, the publications field (for EBI data) or BioProject field (for NCBI data) will be used to indicate the source for each row in the input file, but if this is missing or you want to override it with a single value for all samples, you may provide a source name via this parameter.}
}
\value{
A data frame with the processed AST data, including additional columns:
\itemize{
\item \code{id}: The biosample identifier.
\item \code{spp_pheno}: The species phenotype, formatted using the \code{as.mo} function.
\item \code{drug_agent}: The antibiotic used in the test, formatted using the \code{as.ab} function.
\item \code{mic}: The minimum inhibitory concentration (MIC) value, formatted using the \code{as.mic} function.
\item \code{disk}: The disk diffusion measurement (in mm), formatted using the \code{as.disk} function.
\item \code{method}: The AST platform recorded in the input file as the source of the measurement.
\item \code{guideline}: The AST standard recorded in the input file as being used for the AST assay.
\item \code{pheno_eucast}: The phenotype newly interpreted against EUCAST human breakpoint standards (as S/I/R), based on the MIC or disk diffusion data.
\item \code{pheno_clsi}: The phenotype newly interpreted against CLSI human breakpoint standards (as S/I/R), based on the MIC or disk diffusion data.
\item \code{ecoff}: The phenotype newly interpreted against the ECOFF (as S/R), based on the MIC or disk diffusion data.
\item \code{pheno_provided}: The original phenotype interpretation provided in the input file.
\item \code{source}: The source of each data point (from the publications or bioproject field in the input file, or replaced with a single value passed in as the 'source' parameter).
}
}
\description{
This function imports an antibiotic susceptibility testing (AST) dataset, processes the data, and optionally interprets the results based on MIC or disk diffusion data. It assumes that the input file is a tab-delimited text file (e.g., TSV) or CSV (which may be commpressed) and parses relevant columns (antibiotic names, species names, MIC or disk data) into suitable classes using the AMR package. It optionally can use the AMR package to interpret susceptibility phenotype (SIR) based on EUCAST or CLSI guidelines (human breakpoints and/or ECOFF). If expected columns are not found warnings will be given, and interpretation may not be possible.
}
\examples{
\dontrun{
# small example E. coli AST data from NCBI
ecoli_ast_raw

# import without re-interpreting resistance
pheno <- import_ei_ast(ecoli_ast_raw)
head(pheno)

# import and re-interpret resistance (S/I/R) and WT/NWT (vs ECOFF) using AMR package
pheno <- import_ei_ast(ecoli_ast_raw, interpret_eucast = TRUE, interpret_ecoff = TRUE)
head(pheno)
}
}
