% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pheno.R
\name{import_ncbi_ast}
\alias{import_ncbi_ast}
\title{Import and process AST data from an NCBI file}
\usage{
import_ncbi_ast(file, interpret = T)
}
\arguments{
\item{file}{A string representing the path to the delimited file containing the AST data
in NCBI antibiogram format. These files can be downloaded from the NCBI AST browser,
e.g. https://www.ncbi.nlm.nih.gov/pathogens/ast#Pseudomonas\%20aeruginosa}

\item{interpret}{A logical value (default is TRUE). If \code{TRUE}, the function will interpret
the susceptibility phenotype (SIR) for each row based on the MIC or disk diffusion
values, and either EUCAST or CLSI testing standard (as indicated in the \verb{Testing standard}
column of the input file, if blank EUCAST will be used by default). If \code{FALSE}, no
interpretation is performed.}
}
\value{
A data frame with the processed AST data, including additional columns:
\itemize{
\item \code{biosample}: The biological sample identifier (renamed from \verb{#BioSample}).
\item \code{spp_pheno}: The species phenotype, formatted using the \code{as.mo} function.
\item \code{drug_agent}: The antibiotic used in the test, formatted using the \code{as.ab} function.
\item \code{mic}: The minimum inhibitory concentration (MIC) value, formatted using the \code{as.mic} function.
\item \code{disk}: The disk diffusion measurement (in mm), formatted using the \code{as.disk} function.
\item \code{guideline}: The guideline used for interpretation (either EUCAST or CLSI).
\item \code{pheno}: The interpreted phenotype (SIR) based on the MIC or disk diffusion data.
}
}
\description{
This function imports an antibiotic susceptibility testing (AST) dataset, processes
the data, and optionally interprets the results based on MIC or disk diffusion data.
It assumes that the input file is a delimited text file (e.g., CSV, TSV) and
parses relevant columns (antibiotic names, species names, MIC or disk data) into
suitable classes using the AMR package. It also uses the AMR package to determine
susceptibility phenotype (SIR) based on EUCAST or CLSI guidelines
}
\examples{
# Example usage
ast_data <- import_ncbi_ast("path/to/ast_data.tsv", interpret = TRUE)
head(ast_data)

}
