% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{logistf_details}
\alias{logistf_details}
\title{Extract Details from a logistf Model}
\usage{
logistf_details(model)
}
\arguments{
\item{model}{A fitted logistf model object.}
}
\value{
A tibble containing the estimates, confidence intervals, and p-values for each predictor in the model.
}
\description{
This function extracts and formats the estimates, confidence intervals, and p-values from a fitted logistf model.
}
\examples{

library(logistf)
library(ggplot2)
model <- logistf(case ~ age + oc + vic + vicl + vis + dia, data = sex2)
model_details <- logistf_details(model)
autoplot(model_details)

}
