% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ST_data_CLI}
\alias{ST_data_CLI}
\title{ST data for \emph{Staphylococcus aureus} genomes for clindamycin vignette}
\format{
\code{ST_data_CLI} A data frame with 37934 rows and 10 columns:
\itemize{
\item \code{Name}: Sample identifier, imported from the \verb{#BioSample} column in the raw input.
\item \code{schema}: Typing scheme used to determine sequence type (ST), in this case \code{pubmlst_saureus}.
\item \code{ST}: Sequence type, determined by PubMLST scheme for S. aureus.
\item \verb{arc allele}: Allele identifier for the \code{arc} locus in the PubMLST scheme.
\item \verb{aroE allele}: Allele identifier for the \code{aroE} locus in the PubMLST scheme.
\item \verb{glpF allele}: Allele identifier for the \code{glpF} locus in the PubMLST scheme.
\item \verb{gmk allele}: Allele identifier for the \code{gmk} locus in the PubMLST scheme.
\item \verb{pta allele}: Allele identifier for the \code{pta} locus in the PubMLST scheme.
\item \verb{tpi allele}: Allele identifier for the \code{tpi} locus in the PubMLST scheme.
\item \verb{yqiL allele}: Allele identifier for the \code{yqiL} locus in the PubMLST scheme.
}
}
\source{
\url{https://github.com/AllTheBacteria/AllTheBacteria}
}
\usage{
ST_data_CLI
}
\description{
Used to investigate the distribution of
clindamycin resistance determinants across lineages. Assemblies were generated by the
AllTheBacteria project and STs were determined based one the PubMLST scheme for
S. aureus \url{https://pubmlst.org/bigsdb/page/schemes}, using bactopia run by Robert Petit.
}
\keyword{datasets}
